Pull Philippines data on cases information from the publicly available Department of Health COVID-19 Data Drop

A wrapper to googledrive and googlesheets4 functions to pull data from the Philppines' COVID-19 Data Drop resource that is publicly distributed via Google Drive

datadrop_cases(version = "current", .date = NULL)

Arguments

version
A character value specifying whether to get the most currently available dataset ("current") or to get archive data ("archive"). Default to "current"

.date
A character value for date in YYYY-MM-DD format. This is the date up to which extracted data reports to. Only used when version is set to "archive" otherwise ignored.

Value

A tibble of case information on confirmed COVID-19 cases distributed via the COVID-19 Data Drop

Examples

datadrop_cases()

#> Getting information on Google Drive directory structure of the DoH Data Dropfor latest available data up to 2020-11-11.

#> # A tibble: 401,416 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C636313     43 "40 to … MALE  2020-08-02   "2020-08-03"     2020-08-06 
#>  2 C896468     28 "25 to … MALE  2020-07-15   "2020-07-16"     2020-07-20 
#>  3 C590610     33 "30 to … MALE  2020-05-17   ""               2020-05-31 
#>  4 C562528     56 "55 to … MALE  2020-07-13   "2020-07-14"     2020-07-17 
#>  5 C343867     39 "35 to … FEMA… 2020-07-20   "2020-04-23"     2020-08-01 
#>  6 C339734     37 "35 to … FEMA… 2020-07-25   "2020-07-27"     2020-08-01 
#>  7 C530113     NA ""       FEMA… 2020-07-15   "2020-07-18"     2020-07-21 
#>  8 C850092     71 "70 to … MALE  2020-07-27   "2020-08-01"     2020-08-05 
#>  9 C908957     67 "65 to … MALE  2020-06-25   "2020-06-25"     2020-06-27 
#> 10 C955592     23 "20 to … FEMA… 2020-07-17   "2020-07-19"     2020-07-22 
#> # … with 401,406 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>