Retrieve datasets from specified DoH Data Drop folders

A wrapper to googledrive functions to retrieve datasets from the DoH Data Drop folders.

datadrop_get(id, path = NULL, keep = FALSE, overwrite = FALSE, verbose = TRUE)

datadrop_get_changelog(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  version = c("latest", "archive"),
  .date = NULL,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

id
A 33-character string identifier for the Google Drive file

path
A character value for path for output file. If NULL, the default file name used in Google Drive is used and the default location is the working directory.

keep
Logical. Should file be saved locally? Default to FALSE. If TRUE, file is kept in the location specified in path. If path is NULL, the file is kept in the working directory using the default file name used in Google Drive.

overwrite
Logical. If path already exists, should it be overwritten? Default to FALSE.

verbose
Logical. Should operation progress messages be shown? Default to TRUE.

version
A character value specifying whether to get the latest available dataset (latest) or to get archive data (archive). Default to latest.

.date
A character value for date in YYYY-MM-DD format. This is the date for the archive DoH Data Drop for which an ID is to be returned. Should be specified when using version is set to archive otherwise ignored.

Value

A tibble of retrieved DoH Data Drop file. If keep is TRUE, a file is also downloaded into specified path.

A tibble of any of the following datasets:

Metadata - Sheets;

Metadata - Fields;

Case Information;

DOH Data Collect - Daily Report;

DOH Data Collect - Weekly Report;

Testing Aggregates;

Quarantine Facility Data - Daily Report;

Quarantine Facility Data - Weekly Report;

DOH Data Collect v3 - Baseline;

DOH Data Collect v4 - Baseline; and,

DDC TTMF Tracker v1. For Changelog, a named list of two tibbles - List of Changes and Most Common Changes. If keep is TRUE, a copy of the specified DoH Data Drop Google Drive file is saved in the location specified by path.

Examples

## Get Google Drive ID for latest case information data
id <- datadrop_id_file(tbl = datadrop_ls(id = datadrop_id()),
                       fn = "Case")

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f084030ae67

## Retrieve case information data
datadrop_get(id = id, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f083c0a0d1d

#> # A tibble: 406,337 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C912262     57 55 to 59 MALE  "2020-07-25" "2020-08-01"     2020-08-04 
#>  2 C735296     67 65 to 69 FEMA… "2020-09-09" "2020-09-11"     2020-09-13 
#>  3 C375626     36 35 to 39 MALE  "2020-07-20" "2020-07-25"     2020-08-01 
#>  4 C750303     15 15 to 19 FEMA… "2020-07-27" "2020-08-02"     2020-08-05 
#>  5 C855142     30 30 to 34 FEMA… ""           ""               2020-06-04 
#>  6 C781806     27 25 to 29 FEMA… "2020-07-24" "2020-07-29"     2020-08-02 
#>  7 C819239     28 25 to 29 MALE  ""           "2020-04-24"     2020-04-27 
#>  8 C786885     65 65 to 69 MALE  "2020-07-19" "2020-07-21"     2020-07-25 
#>  9 C800563      9 5 to 9   FEMA… "2020-07-13" "2020-07-16"     2020-07-19 
#> 10 C587000     25 25 to 29 MALE  ""           ""               2020-07-21 
#> # … with 406,327 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest changelog information
datadrop_get_changelog(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f0837e309b6

#> Warning: File/s with the word/s Changelog.xlsx was not found in the Data Drop folder.
#> Please revise as needed and try again. Returning NULL.

#> Warning: No changelog information on 2020-11-14. Try a date earlier or later than date
#> specified. Returning NULL.

#> NULL

## Retrieve latest metadata - sheets information
datadrop_get_sheets(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f081a92a864

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 02 Metadata - Sheets.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f08126d4add

#> # A tibble: 10 x 5
#>    Sheet     Description          Source       Contact.Informa… Data.Update.Sch…
#>    <chr>     <chr>                <chr>        <chr>            <chr>           
#>  1 Case Inf… "List of confirmed … DOH - Epide… covidtracker@do… Daily at 4PM    
#>  2 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  3 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  4 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  5 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  6 DDC TTMF… "Self-reported staf… Collated by… covidtracker@do… Daily at 4PM    
#>  7 DOH Data… "Self-reported coun… Collated by… covidtracker@do… Daily at 4PM    
#>  8 DOH Data… "Self-reported IPC … Collated by… covidtracker@do… Daily at 4PM    
#>  9 Testing … ""                   DOH          covidtracker@do… Daily at 4PM    
#> 10 Privacy … ""                   -            -                -               

## Retrieve latest metadata - fields information
datadrop_get_fields(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f08454dfc9e

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 03 Metadata - Fields.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f083040c9e

#> # A tibble: 199 x 7
#>    Sheet  Field  Description      Type  Question.Format Possible.Respon… Purpose
#>    <chr>  <chr>  <chr>            <chr> <chr>           <chr>            <chr>  
#>  1 Case … CaseC… Random code ass… Text  ""              ""               ""     
#>  2 Case … Age    Age              Numb… ""              ""               ""     
#>  3 Case … AgeGr… Five-year age g… Text  ""              ""               ""     
#>  4 Case … Sex    Sex              Text  ""              ""               ""     
#>  5 Case … DateS… Date when speci… Date  ""              ""               ""     
#>  6 Case … DateR… Date of release… Date  ""              ""               ""     
#>  7 Case … DateR… Date publicly a… Date  ""              ""               ""     
#>  8 Case … DateD… Date died        Date  ""              ""               ""     
#>  9 Case … DateR… Date recovered   Date  ""              ""               ""     
#> 10 Case … Remov… Type of removal… Text  ""              ""               ""     
#> # … with 189 more rows

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f087ed85a9d

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f082e253a7a

#> # A tibble: 406,337 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C912262     57 55 to 59 MALE  "2020-07-25" "2020-08-01"     2020-08-04 
#>  2 C735296     67 65 to 69 FEMA… "2020-09-09" "2020-09-11"     2020-09-13 
#>  3 C375626     36 35 to 39 MALE  "2020-07-20" "2020-07-25"     2020-08-01 
#>  4 C750303     15 15 to 19 FEMA… "2020-07-27" "2020-08-02"     2020-08-05 
#>  5 C855142     30 30 to 34 FEMA… ""           ""               2020-06-04 
#>  6 C781806     27 25 to 29 FEMA… "2020-07-24" "2020-07-29"     2020-08-02 
#>  7 C819239     28 25 to 29 MALE  ""           "2020-04-24"     2020-04-27 
#>  8 C786885     65 65 to 69 MALE  "2020-07-19" "2020-07-21"     2020-07-25 
#>  9 C800563      9 5 to 9   FEMA… "2020-07-13" "2020-07-16"     2020-07-19 
#> 10 C587000     25 25 to 29 MALE  ""           ""               2020-07-21 
#> # … with 406,327 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f08e223741

#> Warning: File/s with the word/s ChangelCollect - Daily Report.csv was not found in the
#> Data Drop folder. Please revise as needed and try again. Returning NULL.

#> Warning: No daily hospital beds and mechanical ventilator information on 2020-11-14. Try
#> a date earlier or later than date specified. Returning NULL.

#> NULL

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f08685993a6

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 06 DOH Data Collect - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f0850ecaccf

#> # A tibble: 53,945 x 26
#>    hfhudcode     id cfname updateddate addeddate reportdate  gown gloves
#>    <chr>      <dbl> <chr>  <chr>       <chr>     <chr>      <dbl>  <dbl>
#>  1 DOH00000…  31175 DUBON… 2020-04-24… 2020-04-… 2020-04-19    10     50
#>  2 DOH00000…  33300 DUBON… 2020-05-01… 2020-05-… 2020-04-26    10     50
#>  3 DOH00000…  34411 DUBON… 2020-05-08… 2020-05-… 2020-05-03    10     50
#>  4 DOH00000…  36024 DUBON… 2020-05-15… 2020-05-… 2020-05-10    10     50
#>  5 DOH00000…  36916 DUBON… 2020-05-22… 2020-05-… 2020-05-17    10     50
#>  6 DOH00000…  39189 DUBON… 2020-05-29… 2020-05-… 2020-05-24    10     50
#>  7 DOH00000… 311478 DUBON… 2020-06-05… 2020-06-… 2020-05-31    10     50
#>  8 DOH00000… 311892 DUBON… 2020-06-12… 2020-06-… 2020-06-07    10     50
#>  9 DOH00000… 314358 DUBON… 2020-06-19… 2020-06-… 2020-06-14    10     50
#> 10 DOH00000… 315386 DUBON… 2020-06-26… 2020-06-… 2020-06-21    10     50
#> # … with 53,935 more rows, and 18 more variables: head_cover <dbl>,
#> #   goggles <dbl>, coverall <dbl>, shoe_cover <dbl>, face_shield <dbl>,
#> #   surgmask <dbl>, n95mask <dbl>, ops <dbl>, nps <dbl>, vtm <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest testing aggregates information
datadrop_get_tests(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f086245738a

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 07 Testing Aggregates.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f083da49b6d

#> # A tibble: 18,969 x 18
#>    facility_name report_date avg_turnaround_… daily_output_sa… daily_output_un…
#>    <chr>         <chr>                  <dbl>            <dbl>            <dbl>
#>  1 AL Molecular… 2020-08-30                24                0                0
#>  2 AL Molecular… 2020-08-31                24                0                0
#>  3 AL Molecular… 2020-09-01                NA               NA               NA
#>  4 AL Molecular… 2020-09-02                NA               NA               NA
#>  5 AL Molecular… 2020-09-03                NA               NA               NA
#>  6 AL Molecular… 2020-09-04                NA               NA               NA
#>  7 AL Molecular… 2020-09-05                NA               NA               NA
#>  8 AL Molecular… 2020-09-06                NA               NA               NA
#>  9 AL Molecular… 2020-09-07                 8                1                1
#> 10 AL Molecular… 2020-09-08                 8                1                1
#> # … with 18,959 more rows, and 13 more variables:
#> #   daily_output_positive_individuals <dbl>,
#> #   daily_output_negative_individuals <dbl>, daily_output_equivocal <dbl>,
#> #   daily_output_invalid <dbl>, remaining_available_tests <dbl>,
#> #   backlogs <dbl>, cumulative_samples_tested <int>,
#> #   cumulative_unique_individuals <int>, cumulative_positive_individuals <int>,
#> #   cumulative_negative_individuals <int>, pct_positive_cumulative <dbl>,
#> #   pct_negative_cumulative <dbl>, validation_status <chr>

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f081d830b9

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 08 Quarantine Facility Data - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f081866c2a1

#> # A tibble: 1,724,486 x 22
#>    updateddate addeddate reportdate other_adm doctor_adm nurse_adm qother
#>    <chr>       <chr>     <chr>          <dbl>      <dbl>     <dbl>  <dbl>
#>  1 2020-06-01… 2020-06-… 2020-06-0…         0          0         0      0
#>  2 2020-06-02… 2020-06-… 2020-06-0…         0          0         0      0
#>  3 2020-06-03… 2020-06-… 2020-06-0…         0          0         0      0
#>  4 2020-06-04… 2020-06-… 2020-06-0…         0          0         0      0
#>  5 2020-06-05… 2020-06-… 2020-06-0…         0          0         0      0
#>  6 2020-06-06… 2020-06-… 2020-06-0…         0          0         0      0
#>  7 2020-06-07… 2020-06-… 2020-06-0…         0          0         0      0
#>  8 2020-06-08… 2020-06-… 2020-06-0…         0          0         0      0
#>  9 2020-06-09… 2020-06-… 2020-06-0…         0          0         0      0
#> 10 2020-06-10… 2020-06-… 2020-06-1…         0          0         0      0
#> # … with 1,724,476 more rows, and 15 more variables: qdoctor <dbl>,
#> #   qnurse <dbl>, isolbed_o <dbl>, isolbed_v <dbl>, beds_ward_o <dbl>,
#> #   beds_ward_v <dbl>, facilityname <chr>, region <chr>, province <chr>,
#> #   city_mun <chr>, bgy <chr>, region_psgc <chr>, province_psgc <chr>,
#> #   city_mun_psgc <chr>, bgy_psgc <chr>

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(path = tempfile())

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-14.

#> File downloaded:
#>   * 01 READ ME FIRST (11_14).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f08273e48b

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201114 - 09 Quarantine Facility Data - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//Rtmp7rldZi/file3f086a162ef

#> # A tibble: 1,387 x 25
#>    facilityname updateddate addeddate reportdate coverall surgmask shoe_cover
#>    <chr>        <chr>       <chr>     <chr>         <dbl>    <dbl>      <dbl>
#>  1 ACABAN ELEM… ""          ""        2020-06-0…       NA       NA         NA
#>  2 CASSILY LAK… ""          ""        2020-05-2…       NA       NA         NA
#>  3 CASSILY LAK… "2020-06-0… "2020-06… 2020-05-3…       10       10         10
#>  4 CASSILY LAK… ""          ""        2020-06-0…       NA       NA         NA
#>  5 BALER MUNIC… "2020-05-2… "2020-05… 2020-05-2…       15      100         15
#>  6 BALER MUNIC… "2020-06-0… "2020-06… 2020-05-3…       15      100         15
#>  7 BALER MUNIC… ""          ""        2020-06-0…       NA       NA         NA
#>  8 CAWAYAN BAR… ""          ""        2020-06-0…       NA       NA         NA
#>  9 LOURDES NAT… ""          ""        2020-06-0…       NA       NA         NA
#> 10 MACROHON CIU "2020-06-0… "2020-06… 2020-05-3…       36    16250         34
#> # … with 1,377 more rows, and 18 more variables: face_shield <dbl>,
#> #   head_cover <dbl>, n95mask <dbl>, goggles <dbl>, gloves <dbl>, gown <dbl>,
#> #   vtm <dbl>, nps <dbl>, ops <dbl>, weekly_reportdate <chr>, region <chr>,
#> #   province <chr>, city_mun <chr>, bgy <chr>, region_psgc <chr>,
#> #   province_psgc <chr>, city_mun_psgc <chr>, bgy_psgc <chr>

#datadrop_get_collectV3(path = tempfile())
#datadrop_get_collectV4(path = tempfile())
#datadrop_get_tracker(path = tempfile())