Get Google Drive IDs for latest or archive DoH Data Drop folders

The Philippines Department of Health currently distributes the latest data drop via a fixed shortened URL (bit.ly/DataDropPH) which links/points to a new Google Drive endpoint daily or whenever the daily updated data drop is available. This Google Drive endpoint is a README document in portable document format (PDF) which contains a privacy and confidentiality statement, technical notes with regard to the latest data, technical notes with regard to previous (archive data) and two shortened URLs - one linking to the Google Drive folder that contains all the latest officially released datasets, and the other linking to the datasets released previously (archives). Of these, the first shortened URL linking to the Google Drive folder containing the latest officially released datasets is different for everyday and can only be obtained through the README document released for a specific day.

datadrop_latest_id(verbose = TRUE)

datadrop_archive_id(verbose = TRUE, .date = NULL)

datadrop_id(verbose = TRUE, version = c("latest", "archive"), .date = NULL)

Arguments

verbose
Logical. Should message on operation progress be shown. Default is TRUE.

.date
A character value for date in YYYY-MM-DD format. This is the date for the archive DoH Data Drop for which an ID is to be returned. Should be specified when using datadrop_archive_id(). For datadrop_id(), only used when version is set to archive otherwise ignored.

version
A character value specifying whether to get the latest available DoH Data Drop (latest) or to get DoH Data Drop archive (archive). Default to latest.

Value

A 33-character string for the Google Drive ID of the latest DoH Data Drop or the archive DoH Data Drop

Details

The function datadrop_latest_id() reads that PDF file, extracts the shortened URL for the latest official released datasets written in that file, expands that shortened URL and then extracts the unique Google Drive ID for the latest officially released datasets. With this Google Drive ID, other functions can then be used to retrieve information and data from the Google Drive specified by this ID.

The DoH Data Drop archives, on the other hand, is distributed via a fixed shortened URL (bit.ly/DataDropArchives) which links/points to a Google Drive folder containing the previous Data Drop releases.

The function datadrop_archive_id() expands that shortened URL and then extracts the unique Google Drive ID for the DoH Data Drop archives folder. With this Google Drive ID, other functions can then be used to retrieve information and data from the Google Drive specified by this ID.

Author

Ernest Guevarra

Examples

datadrop_latest_id()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> [1] "1ldBJ8dvz-WvxiB7-yLCbEKaLuuWL6t6q"

datadrop_archive_id(.date = "2020-11-01")

#> Getting the Google Drive ID for the DoH Data Drop archive for 2020-11-01

#> [1] "1O2Gt_MUPKtWWPK6ainagRiehK0pCH7Nv"

datadrop_id()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> [1] "1ldBJ8dvz-WvxiB7-yLCbEKaLuuWL6t6q"

datadrop_id(version = "archive", .date = "2020-11-01")

#> Getting the Google Drive ID for the DoH Data Drop archive for 2020-11-01

#> [1] "1O2Gt_MUPKtWWPK6ainagRiehK0pCH7Nv"