Get Philippines Department of Health COVID-19 Data Drop Google Drive specific directory information

A wrapper to googledrive functions to extract information on contents of a specific COVID-19 Data Drop Google Drive directory

datadrop_ls(version = "current", .date = NULL)

Arguments

version
A character value specifying whether to get the most currently available dataset ("current") or to get archive data ("archive"). Default to "current"

.date
A character value for date in YYYY-MM-DD format. This is the date up to which extracted data reports to. Only used when version is set to "archive" otherwise ignored.

Value

A tibble containing information on the various files in a specified COVID-19 Data Drop Google Drive directory

Examples

datadrop_ls()

#> Getting information on Google Drive directory structure of the DoH Data Drop for latest available data up to 2020-11-12.

#> # A tibble: 13 x 3
#>    name                                      id                  drive_resource 
#>  * <chr>                                     <chr>               <list>         
#>  1 DOH COVID Data Drop_ 20201112 - 07 Testi… 1t4embbR8SpOm12dGD… <named list [3…
#>  2 DOH COVID Data Drop_ 20201112 - 07 Testi… 1yffN3ofeQx3iw944u… <named list [3…
#>  3 DOH COVID Data Drop_ 20201112 - 04 Case … 1IsHcntUetrTPIl9sI… <named list [3…
#>  4 DOH COVID Data Drop_ 20201112 - 08 Quara… 1AEwisdKOhWKS8KDdj… <named list [3…
#>  5 DOH COVID Data Drop_ 20201112 - 05 DOH D… 1lXMwI3BXp7qN0W-Eq… <named list [3…
#>  6 DOH COVID Data Drop_ 20201112 - 09 Quara… 1UKjpu6a0M0nt1FN_U… <named list [3…
#>  7 DOH COVID Data Drop_ 20201112 - 02 Metad… 12OUSL0q_zIf-wMDmM… <named list [3…
#>  8 DOH COVID Data Drop_ 20201112 - 06 DOH D… 1DlW35JFoCEBQ-NY0a… <named list [3…
#>  9 DOH COVID Data Drop_ 20201112 - 11 DOH D… 18SokHdLJUUbq1F2nm… <named list [3…
#> 10 DOH COVID Data Drop_ 20201112 - 12 DDC T… 1Jkmc3RpSHcciLSgdX… <named list [3…
#> 11 DOH COVID Data Drop_ 20201112 - 10 DOH D… 1uE5bkJ81gT6vbvTI7… <named list [3…
#> 12 DOH COVID Data Drop_ 20201112 - 03 Metad… 1yb022ZgJv3kE9GM-U… <named list [3…
#> 13 01 READ ME FIRST (11_12).pdf              1YFHhWZNshEj2E9ssa… <named list [3…