Retrieve datasets from specified DoH Data Drop folders

A wrapper to googledrive functions to retrieve datasets from the DoH Data Drop folders.

datadrop_get(
  tbl,
  fn,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_changelog(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

tbl
A tibble output produced by datadrop_ls() that lists the files within a particular DoH Data Drop Google Drive folder

fn
A character string composed of a word or words that can be used to match to the name of a file within a particular DoH Data Drop Google Drive folder listed in tbl.

path
A character value for path for output file. If NULL, the default file name used in Google Drive is used and the default location is the working directory.

keep
Logical. Should file be saved locally? Default to FALSE. If TRUE, file is kept in the location specified in path. If path is NULL, the file is kept in the working directory using the default file name used in Google Drive.

overwrite
Logical. If path already exists, should it be overwritten? Default to FALSE.

verbose
Logical. Should operation progress messages be shown? Default to TRUE.

Value

A tibble of any of the following datasets:

Metadata - Sheets;

Metadata - Fields;

Case Information;

DOH Data Collect - Daily Report;

DOH Data Collect - Weekly Report;

Testing Aggregates;

Quarantine Facility Data - Daily Report;

Quarantine Facility Data - Weekly Report;

DOH Data Collect v3 - Baseline;

DOH Data Collect v4 - Baseline; and,

DDC TTMF Tracker v1. For Changelog, a named list of two tibbles - List of Changes and Most Common Changes. If keep is TRUE, a copy of the specified DoH Data Drop Google Drive file is saved in the location specified by path.

Examples

## Get tbl for files in latest Data Drop
library(magrittr)
x <- datadrop_id() %>% datadrop_ls()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-21.

#> File downloaded:
#>   * 01 READ ME FIRST (11_21).pdf
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc7524408f85

## Retrieve case information data
datadrop_get(tbl = x, fn = "Case Information", path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc75a8e775b

#> # A tibble: 416,852 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C581144     53 50 to 54 MALE  "2020-07-15" "2020-07-16"     2020-07-21 
#>  2 C352619     56 55 to 59 FEMA… ""           ""               2020-05-04 
#>  3 C466590     36 35 to 39 MALE  "2020-07-14" "2020-07-17"     2020-07-20 
#>  4 C971812     44 40 to 44 MALE  "2020-05-07" ""               2020-05-21 
#>  5 C445586     57 55 to 59 MALE  "2020-07-14" "2020-07-16"     2020-07-19 
#>  6 C354376     34 30 to 34 FEMA… "2020-04-14" "2020-04-16"     2020-09-01 
#>  7 C583142     54 50 to 54 MALE  "2020-07-20" "2020-07-25"     2020-08-02 
#>  8 C329630     56 55 to 59 FEMA… "2020-07-20" "2020-07-22"     2020-07-27 
#>  9 C349878     22 20 to 24 MALE  "2020-07-22" "2020-07-26"     2020-07-30 
#> 10 C956795     48 45 to 49 FEMA… "2020-07-30" "2020-08-01"     2020-08-04 
#> # … with 416,842 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest changelog information
datadrop_get_changelog(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH Data Drop 20201121 - Changelog.xlsx
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc75f3dfcc7

#> $`List of Changes`
#> # A tibble: 6,539 x 4
#>    CaseCode Column       `Old Value` `New Value` 
#>    <chr>    <chr>        <chr>       <chr>       
#>  1 C100017  HealthStatus MILD        ASYMPTOMATIC
#>  2 C100017  Quarantined  NO          YES         
#>  3 C100350  HealthStatus MILD        RECOVERED   
#>  4 C100350  RemovalType  NA          RECOVERED   
#>  5 C100387  DateRecover  NA          44129.0     
#>  6 C100417  HealthStatus RECOVERED   DIED        
#>  7 C100417  DateRecover  44089.0     NA          
#>  8 C100417  DateDied     NA          44073.0     
#>  9 C100417  RemovalType  RECOVERED   DIED        
#> 10 C100992  DateSpecimen 44036.0     NA          
#> # … with 6,529 more rows
#> 
#> $`Most Common Changes`
#> # A tibble: 1,809 x 4
#>    Column       `Old Value`  `New Value`  `Records Changed`
#>    <chr>        <chr>        <chr>                    <dbl>
#>  1 Quarantined  NO           YES                        621
#>  2 HealthStatus MILD         ASYMPTOMATIC               326
#>  3 RemovalType  NA           RECOVERED                  309
#>  4 HealthStatus MILD         RECOVERED                  212
#>  5 Admitted     NO           YES                        120
#>  6 HealthStatus ASYMPTOMATIC RECOVERED                   80
#>  7 DateRecover  NA           44133.0                     78
#>  8 DateRecover  NA           44134.0                     76
#>  9 DateRecover  NA           44136.0                     57
#> 10 DateRecover  NA           44154.0                     56
#> # … with 1,799 more rows
#> 

## Retrieve latest metadata - sheets information
datadrop_get_sheets(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 02 Metadata - Sheets.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc752e9e74a2

#> # A tibble: 10 x 5
#>    Sheet     Description          Source       Contact.Informa… Data.Update.Sch…
#>    <chr>     <chr>                <chr>        <chr>            <chr>           
#>  1 Case Inf… "List of confirmed … DOH - Epide… covidtracker@do… Daily at 4PM    
#>  2 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  3 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  4 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  5 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  6 DDC TTMF… "Self-reported staf… Collated by… covidtracker@do… Daily at 4PM    
#>  7 DOH Data… "Self-reported coun… Collated by… covidtracker@do… Daily at 4PM    
#>  8 DOH Data… "Self-reported IPC … Collated by… covidtracker@do… Daily at 4PM    
#>  9 Testing … ""                   DOH          covidtracker@do… Daily at 4PM    
#> 10 Privacy … ""                   -            -                -               

## Retrieve latest metadata - fields information
datadrop_get_fields(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 03 Metadata - Fields.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc7524fb4f97

#> # A tibble: 199 x 7
#>    Sheet  Field  Description      Type  Question.Format Possible.Respon… Purpose
#>    <chr>  <chr>  <chr>            <chr> <chr>           <chr>            <chr>  
#>  1 Case … CaseC… Random code ass… Text  ""              ""               ""     
#>  2 Case … Age    Age              Numb… ""              ""               ""     
#>  3 Case … AgeGr… Five-year age g… Text  ""              ""               ""     
#>  4 Case … Sex    Sex              Text  ""              ""               ""     
#>  5 Case … DateS… Date when speci… Date  ""              ""               ""     
#>  6 Case … DateR… Date of release… Date  ""              ""               ""     
#>  7 Case … DateR… Date publicly a… Date  ""              ""               ""     
#>  8 Case … DateD… Date died        Date  ""              ""               ""     
#>  9 Case … DateR… Date recovered   Date  ""              ""               ""     
#> 10 Case … Remov… Type of removal… Text  ""              ""               ""     
#> # … with 189 more rows

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc756f265578

#> # A tibble: 416,852 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C581144     53 50 to 54 MALE  "2020-07-15" "2020-07-16"     2020-07-21 
#>  2 C352619     56 55 to 59 FEMA… ""           ""               2020-05-04 
#>  3 C466590     36 35 to 39 MALE  "2020-07-14" "2020-07-17"     2020-07-20 
#>  4 C971812     44 40 to 44 MALE  "2020-05-07" ""               2020-05-21 
#>  5 C445586     57 55 to 59 MALE  "2020-07-14" "2020-07-16"     2020-07-19 
#>  6 C354376     34 30 to 34 FEMA… "2020-04-14" "2020-04-16"     2020-09-01 
#>  7 C583142     54 50 to 54 MALE  "2020-07-20" "2020-07-25"     2020-08-02 
#>  8 C329630     56 55 to 59 FEMA… "2020-07-20" "2020-07-22"     2020-07-27 
#>  9 C349878     22 20 to 24 MALE  "2020-07-22" "2020-07-26"     2020-07-30 
#> 10 C956795     48 45 to 49 FEMA… "2020-07-30" "2020-08-01"     2020-08-04 
#> # … with 416,842 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 05 DOH Data Collect - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc753db5724a

#> # A tibble: 179,925 x 54
#>    hfhudcode    id cfname updateddate addeddate reportdate icu_v icu_o isolbed_v
#>    <chr>     <dbl> <chr>  <chr>       <chr>     <chr>      <dbl> <dbl>     <dbl>
#>  1 DOH00000…  2461 GALIN… 2020-07-03… 2020-07-… 2020-07-0…     0     0         0
#>  2 DOH00000…  3195 GALIN… 2020-07-04… 2020-07-… 2020-07-0…     0     0         0
#>  3 DOH00000…  4491 GALIN… 2020-07-05… 2020-07-… 2020-07-0…     0     0         0
#>  4 DOH00000…  6489 GALIN… 2020-07-06… 2020-07-… 2020-07-0…     0     0         0
#>  5 DOH00000…  8655 GALIN… 2020-07-07… 2020-07-… 2020-07-0…     0     0         0
#>  6 DOH00000…  9533 GALIN… 2020-07-08… 2020-07-… 2020-07-0…     0     0         0
#>  7 DOH00000… 11761 GALIN… 2020-07-09… 2020-07-… 2020-07-0…     0     0         0
#>  8 DOH00000… 13445 GALIN… 2020-07-10… 2020-07-… 2020-07-1…     0     0         0
#>  9 DOH00000… 14720 GALIN… 2020-07-11… 2020-07-… 2020-07-1…     0     0         0
#> 10 DOH00000… 17471 GALIN… 2020-07-12… 2020-07-… 2020-07-1…     0     0         0
#> # … with 179,915 more rows, and 45 more variables: isolbed_o <dbl>,
#> #   beds_ward_v <dbl>, beds_ward_o <dbl>, mechvent_v <dbl>, mechvent_o <dbl>,
#> #   icu_v_nc <dbl>, icu_o_nc <dbl>, nonicu_v_nc <dbl>, nonicu_o_nc <dbl>,
#> #   mechvent_v_nc <dbl>, mechvent_o_nc <dbl>, qnurse <dbl>, qdoctor <dbl>,
#> #   qother <dbl>, nurse_adm <dbl>, doctor_adm <dbl>, other_adm <dbl>,
#> #   susp_asym <dbl>, susp_mild <dbl>, susp_severe <dbl>, susp_crit <dbl>,
#> #   susp_died <dbl>, prob_asym <dbl>, prob_mild <dbl>, prob_severe <dbl>,
#> #   prob_crit <dbl>, prob_died <dbl>, conf_asym <dbl>, conf_mild <dbl>,
#> #   conf_severe <dbl>, conf_crit <dbl>, conf_died <dbl>, tpatient_adm <dbl>,
#> #   tpatient_er <dbl>, tpatient_icu <dbl>, trans_ttmf <dbl>, discharged <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 06 DOH Data Collect - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc75536277ec

#> # A tibble: 55,844 x 26
#>    hfhudcode     id cfname updateddate addeddate reportdate  gown gloves
#>    <chr>      <dbl> <chr>  <chr>       <chr>     <chr>      <dbl>  <dbl>
#>  1 DOH00000…  31907 BASIL… "2020-04-2… "2020-04… 2020-04-19    12   2000
#>  2 DOH00000…  32367 BASIL… "2020-05-0… "2020-05… 2020-04-26    12   2000
#>  3 DOH00000…  33841 BASIL… "2020-05-0… "2020-05… 2020-05-03    12   2000
#>  4 DOH00000…     NA BASIL… ""          ""        2020-05-10    NA     NA
#>  5 DOH00000…  37859 BASIL… "2020-05-2… "2020-05… 2020-05-17    12   1500
#>  6 DOH00000…  38698 BASIL… "2020-05-2… "2020-05… 2020-05-24    12   1000
#>  7 DOH00000…     NA BASIL… ""          ""        2020-05-31    NA     NA
#>  8 DOH00000… 311663 BASIL… "2020-06-0… "2020-06… 2020-06-07    12   1000
#>  9 DOH00000… 313786 BASIL… "2020-06-1… "2020-06… 2020-06-14    12    800
#> 10 DOH00000… 315125 BASIL… "2020-06-2… "2020-06… 2020-06-21    12    700
#> # … with 55,834 more rows, and 18 more variables: head_cover <dbl>,
#> #   goggles <dbl>, coverall <dbl>, shoe_cover <dbl>, face_shield <dbl>,
#> #   surgmask <dbl>, n95mask <dbl>, ops <dbl>, nps <dbl>, vtm <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest testing aggregates information
datadrop_get_tests(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 07 Testing Aggregates.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc7565af51b8

#> # A tibble: 20,115 x 18
#>    facility_name report_date avg_turnaround_… daily_output_sa… daily_output_un…
#>    <chr>         <chr>                  <dbl>            <dbl>            <dbl>
#>  1 AL Molecular… 2020-08-30                24                0                0
#>  2 AL Molecular… 2020-08-31                24                0                0
#>  3 AL Molecular… 2020-09-01                NA               NA               NA
#>  4 AL Molecular… 2020-09-02                NA               NA               NA
#>  5 AL Molecular… 2020-09-03                NA               NA               NA
#>  6 AL Molecular… 2020-09-04                NA               NA               NA
#>  7 AL Molecular… 2020-09-05                NA               NA               NA
#>  8 AL Molecular… 2020-09-06                NA               NA               NA
#>  9 AL Molecular… 2020-09-07                 8                1                1
#> 10 AL Molecular… 2020-09-08                 8                1                1
#> # … with 20,105 more rows, and 13 more variables:
#> #   daily_output_positive_individuals <dbl>,
#> #   daily_output_negative_individuals <dbl>, daily_output_equivocal <dbl>,
#> #   daily_output_invalid <dbl>, remaining_available_tests <dbl>,
#> #   backlogs <dbl>, cumulative_samples_tested <int>,
#> #   cumulative_unique_individuals <int>, cumulative_positive_individuals <int>,
#> #   cumulative_negative_individuals <int>, pct_positive_cumulative <dbl>,
#> #   pct_negative_cumulative <dbl>, validation_status <chr>

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 08 Quarantine Facility Data - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc75591e3b2f

#> # A tibble: 1,819,336 x 22
#>    updateddate addeddate reportdate other_adm doctor_adm nurse_adm qother
#>    <chr>       <chr>     <chr>          <dbl>      <dbl>     <dbl>  <dbl>
#>  1 2020-08-29… 2020-08-… 2020-08-2…         0          0         0      0
#>  2 2020-08-30… 2020-08-… 2020-08-3…         0          0         0      0
#>  3 2020-08-31… 2020-08-… 2020-08-3…         0          0         0      0
#>  4 2020-09-01… 2020-09-… 2020-09-0…         0          0         0      0
#>  5 2020-09-02… 2020-09-… 2020-09-0…         0          0         0      0
#>  6 2020-09-03… 2020-09-… 2020-09-0…         0          0         0      0
#>  7 2020-09-04… 2020-09-… 2020-09-0…         0          0         0      0
#>  8 2020-09-05… 2020-09-… 2020-09-0…         0          0         0      0
#>  9 2020-09-06… 2020-09-… 2020-09-0…         0          0         0      0
#> 10 2020-09-07… 2020-09-… 2020-09-0…         0          0         0      0
#> # … with 1,819,326 more rows, and 15 more variables: qdoctor <dbl>,
#> #   qnurse <dbl>, isolbed_o <dbl>, isolbed_v <dbl>, beds_ward_o <dbl>,
#> #   beds_ward_v <dbl>, facilityname <chr>, region <chr>, province <chr>,
#> #   city_mun <chr>, bgy <chr>, region_psgc <chr>, province_psgc <chr>,
#> #   city_mun_psgc <chr>, bgy_psgc <chr>

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201121 - 09 Quarantine Facility Data - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpIEboI1/filefc754fbfb85e

#> # A tibble: 1,382 x 25
#>    facilityname updateddate addeddate reportdate coverall surgmask shoe_cover
#>    <chr>        <chr>       <chr>     <chr>         <dbl>    <dbl>      <dbl>
#>  1 BASIAO ISOL… ""          ""        2020-06-0…       NA       NA         NA
#>  2 RHU MAYORGA… "2020-06-0… "2020-06… 2020-06-0…       72      750        100
#>  3 CASUBIFAN F… "2020-06-0… "2020-06… 2020-05-3…        0        0          0
#>  4 CASUBIFAN F… ""          ""        2020-06-0…       NA       NA         NA
#>  5 CEBU TECHNO… ""          ""        2020-06-0…       NA       NA         NA
#>  6 MUNICIPAL I… ""          ""        2020-06-0…       NA       NA         NA
#>  7 NEW CALAPE … ""          ""        2020-06-0…       NA       NA         NA
#>  8 Pangi Eleme… ""          ""        2020-06-0…       NA       NA         NA
#>  9 BARANGAY AB… "2020-05-2… "2020-05… 2020-05-2…        0        0          0
#> 10 BARANGAY AB… "2020-06-0… "2020-06… 2020-05-3…        0        0          0
#> # … with 1,372 more rows, and 18 more variables: face_shield <dbl>,
#> #   head_cover <dbl>, n95mask <dbl>, goggles <dbl>, gloves <dbl>, gown <dbl>,
#> #   vtm <dbl>, nps <dbl>, ops <dbl>, weekly_reportdate <chr>, region <chr>,
#> #   province <chr>, city_mun <chr>, bgy <chr>, region_psgc <chr>,
#> #   province_psgc <chr>, city_mun_psgc <chr>, bgy_psgc <chr>

#datadrop_get_collectV3(tbl = x, path = tempfile())
#datadrop_get_collectV4(tbl = x, path = tempfile())
#datadrop_get_tracker(tbl = x, path = tempfile())