Retrieve datasets from specified DoH Data Drop folders

A wrapper to googledrive functions to retrieve datasets from the DoH Data Drop folders.

datadrop_get(
  tbl,
  fn,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_changelog(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

tbl
A tibble output produced by datadrop_ls() that lists the files within a particular DoH Data Drop Google Drive folder

fn
A character string composed of a word or words that can be used to match to the name of a file within a particular DoH Data Drop Google Drive folder listed in tbl.

path
A character value for path for output file. If NULL, the default file name used in Google Drive is used and the default location is the working directory.

keep
Logical. Should file be saved locally? Default to FALSE. If TRUE, file is kept in the location specified in path. If path is NULL, the file is kept in the working directory using the default file name used in Google Drive.

overwrite
Logical. If path already exists, should it be overwritten? Default to FALSE.

verbose
Logical. Should operation progress messages be shown? Default to TRUE.

Value

A tibble of any of the following datasets:

Metadata - Sheets;

Metadata - Fields;

Case Information;

DOH Data Collect - Daily Report;

DOH Data Collect - Weekly Report;

Testing Aggregates;

Quarantine Facility Data - Daily Report;

Quarantine Facility Data - Weekly Report;

DOH Data Collect v3 - Baseline;

DOH Data Collect v4 - Baseline; and,

DDC TTMF Tracker v1. For Changelog, a named list of two tibbles - List of Changes and Most Common Changes. If keep is TRUE, a copy of the specified DoH Data Drop Google Drive file is saved in the location specified by path.

Examples

## Get tbl for files in latest Data Drop
library(magrittr)
x <- datadrop_id() %>% datadrop_ls()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-18.

#> File downloaded:
#>   * 01 READ ME FIRST (11_18).pdf
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec540fb4d72a

## Retrieve case information data
datadrop_get(tbl = x, fn = "Case Information", path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec540299a0e74

#> # A tibble: 412,097 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C516975     51 50 to 54 MALE  2020-09-03   2020-09-03       2020-09-05 
#>  2 C643966     25 25 to 29 MALE  2020-08-28   2020-09-09       2020-09-10 
#>  3 C835097     23 20 to 24 FEMA… 2020-08-15   2020-08-17       2020-08-20 
#>  4 C176417     73 70 to 74 MALE  2020-08-13   2020-08-17       2020-08-19 
#>  5 C940421     22 20 to 24 MALE  2020-06-24   2020-06-29       2020-07-04 
#>  6 C650556     26 25 to 29 MALE  2020-10-25   2020-10-28       2020-10-31 
#>  7 C549944     32 30 to 34 MALE  2020-06-19   2020-06-23       2020-06-26 
#>  8 C887500     25 25 to 29 FEMA… 2020-06-24   2020-06-29       2020-07-04 
#>  9 C352799     41 40 to 44 FEMA… 2020-09-27   2020-09-29       2020-10-01 
#> 10 C776384     49 45 to 49 FEMA… 2020-09-14   2020-09-15       2020-09-17 
#> # … with 412,087 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest changelog information
datadrop_get_changelog(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH Data Drop 20201118 - Changelog.xlsx
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec540412af702

#> $`List of Changes`
#> # A tibble: 5,308 x 4
#>    CaseCode Column            `Old Value`   `New Value` 
#>    <chr>    <chr>             <chr>         <chr>       
#>  1 C100027  DateRecover       NA            44124.0     
#>  2 C100048  DateResultRelease 44142.0       NA          
#>  3 C100048  DateSpecimen      44140.0       NA          
#>  4 C100286  DateRecover       NA            44129.0     
#>  5 C100656  BarangayRes       SAN VICENTE I BANABA CERCA
#>  6 C100656  BarangayPSGC      PH042118043   PH042110003 
#>  7 C100656  CityMuniPSGC      PH042118000   PH042110000 
#>  8 C100656  CityMunRes        SILANG        INDANG      
#>  9 C101082  DateDied          NA            44149.0     
#> 10 C101082  HealthStatus      MILD          DIED        
#> # … with 5,298 more rows
#> 
#> $`Most Common Changes`
#> # A tibble: 1,367 x 4
#>    Column       `Old Value` `New Value`           `Records Changed`
#>    <chr>        <chr>       <chr>                             <dbl>
#>  1 RegionRes    NA          Region IV-B: MIMAROPA               336
#>  2 Quarantined  NO          YES                                 287
#>  3 HealthStatus MILD        ASYMPTOMATIC                        160
#>  4 RemovalType  NA          RECOVERED                           127
#>  5 DateOnset    NA          44133.0                             100
#>  6 DateOnset    NA          44134.0                              91
#>  7 DateRecover  NA          44127.0                              81
#>  8 Admitted     NO          YES                                  77
#>  9 RemovalType  NA          DIED                                 77
#> 10 HealthStatus MILD        DIED                                 68
#> # … with 1,357 more rows
#> 

## Retrieve latest metadata - sheets information
datadrop_get_sheets(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 02 Metadata - Sheets.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec5406bbec5ba

#> # A tibble: 10 x 5
#>    Sheet     Description          Source       Contact.Informa… Data.Update.Sch…
#>    <chr>     <chr>                <chr>        <chr>            <chr>           
#>  1 Case Inf… "List of confirmed … DOH - Epide… covidtracker@do… Daily at 4PM    
#>  2 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  3 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  4 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  5 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  6 DDC TTMF… "Self-reported staf… Collated by… covidtracker@do… Daily at 4PM    
#>  7 DOH Data… "Self-reported coun… Collated by… covidtracker@do… Daily at 4PM    
#>  8 DOH Data… "Self-reported IPC … Collated by… covidtracker@do… Daily at 4PM    
#>  9 Testing … ""                   DOH          covidtracker@do… Daily at 4PM    
#> 10 Privacy … ""                   -            -                -               

## Retrieve latest metadata - fields information
datadrop_get_fields(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 03 Metadata - Fields.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec54039a76d99

#> # A tibble: 199 x 7
#>    Sheet  Field  Description      Type  Question.Format Possible.Respon… Purpose
#>    <chr>  <chr>  <chr>            <chr> <chr>           <chr>            <chr>  
#>  1 Case … CaseC… Random code ass… Text  ""              ""               ""     
#>  2 Case … Age    Age              Numb… ""              ""               ""     
#>  3 Case … AgeGr… Five-year age g… Text  ""              ""               ""     
#>  4 Case … Sex    Sex              Text  ""              ""               ""     
#>  5 Case … DateS… Date when speci… Date  ""              ""               ""     
#>  6 Case … DateR… Date of release… Date  ""              ""               ""     
#>  7 Case … DateR… Date publicly a… Date  ""              ""               ""     
#>  8 Case … DateD… Date died        Date  ""              ""               ""     
#>  9 Case … DateR… Date recovered   Date  ""              ""               ""     
#> 10 Case … Remov… Type of removal… Text  ""              ""               ""     
#> # … with 189 more rows

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec5401f0c7561

#> # A tibble: 412,097 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C516975     51 50 to 54 MALE  2020-09-03   2020-09-03       2020-09-05 
#>  2 C643966     25 25 to 29 MALE  2020-08-28   2020-09-09       2020-09-10 
#>  3 C835097     23 20 to 24 FEMA… 2020-08-15   2020-08-17       2020-08-20 
#>  4 C176417     73 70 to 74 MALE  2020-08-13   2020-08-17       2020-08-19 
#>  5 C940421     22 20 to 24 MALE  2020-06-24   2020-06-29       2020-07-04 
#>  6 C650556     26 25 to 29 MALE  2020-10-25   2020-10-28       2020-10-31 
#>  7 C549944     32 30 to 34 MALE  2020-06-19   2020-06-23       2020-06-26 
#>  8 C887500     25 25 to 29 FEMA… 2020-06-24   2020-06-29       2020-07-04 
#>  9 C352799     41 40 to 44 FEMA… 2020-09-27   2020-09-29       2020-10-01 
#> 10 C776384     49 45 to 49 FEMA… 2020-09-14   2020-09-15       2020-09-17 
#> # … with 412,087 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 05 DOH Data Collect - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec5406aee4333

#> # A tibble: 176,082 x 54
#>    hfhudcode    id cfname updateddate addeddate reportdate icu_v icu_o isolbed_v
#>    <chr>     <dbl> <chr>  <chr>       <chr>     <chr>      <dbl> <dbl>     <dbl>
#>  1 DOH00000…   677 MABAM… 2020-07-02… 2020-07-… 2020-07-0…     0     0         1
#>  2 DOH00000…  2296 MABAM… 2020-07-03… 2020-07-… 2020-07-0…     0     0         1
#>  3 DOH00000…  3946 MABAM… 2020-07-04… 2020-07-… 2020-07-0…     0     0         1
#>  4 DOH00000…  3946 MABAM… 2020-07-04… 2020-07-… 2020-07-0…     0     0         1
#>  5 DOH00000…  7211 MABAM… 2020-07-06… 2020-07-… 2020-07-0…     0     0         0
#>  6 DOH00000…  8975 MABAM… 2020-07-07… 2020-07-… 2020-07-0…     0     0         1
#>  7 DOH00000… 10766 MABAM… 2020-07-08… 2020-07-… 2020-07-0…     0     0         1
#>  8 DOH00000… 12724 MABAM… 2020-07-09… 2020-07-… 2020-07-0…     0     0         1
#>  9 DOH00000… 14421 MABAM… 2020-07-10… 2020-07-… 2020-07-1…     0     0         1
#> 10 DOH00000… 16064 MABAM… 2020-07-11… 2020-07-… 2020-07-1…     0     0         1
#> # … with 176,072 more rows, and 45 more variables: isolbed_o <dbl>,
#> #   beds_ward_v <dbl>, beds_ward_o <dbl>, mechvent_v <dbl>, mechvent_o <dbl>,
#> #   icu_v_nc <dbl>, icu_o_nc <dbl>, nonicu_v_nc <dbl>, nonicu_o_nc <dbl>,
#> #   mechvent_v_nc <dbl>, mechvent_o_nc <dbl>, qnurse <dbl>, qdoctor <dbl>,
#> #   qother <dbl>, nurse_adm <dbl>, doctor_adm <dbl>, other_adm <dbl>,
#> #   susp_asym <dbl>, susp_mild <dbl>, susp_severe <dbl>, susp_crit <dbl>,
#> #   susp_died <dbl>, prob_asym <dbl>, prob_mild <dbl>, prob_severe <dbl>,
#> #   prob_crit <dbl>, prob_died <dbl>, conf_asym <dbl>, conf_mild <dbl>,
#> #   conf_severe <dbl>, conf_crit <dbl>, conf_died <dbl>, tpatient_adm <dbl>,
#> #   tpatient_er <dbl>, tpatient_icu <dbl>, trans_ttmf <dbl>, discharged <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 06 DOH Data Collect - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec540407e001d

#> # A tibble: 55,844 x 26
#>    hfhudcode     id cfname updateddate addeddate reportdate  gown gloves
#>    <chr>      <dbl> <chr>  <chr>       <chr>     <chr>      <dbl>  <dbl>
#>  1 DOH00000…   3746 RICO'… 2020-04-24… 2020-04-… 2020-04-19     9    300
#>  2 DOH00000…  32046 RICO'… 2020-05-01… 2020-05-… 2020-04-26     9    300
#>  3 DOH00000…  33518 RICO'… 2020-05-08… 2020-05-… 2020-05-03     9    300
#>  4 DOH00000…  35409 RICO'… 2020-05-15… 2020-05-… 2020-05-10     9    300
#>  5 DOH00000…  36780 RICO'… 2020-05-22… 2020-05-… 2020-05-17     9    300
#>  6 DOH00000…  38385 RICO'… 2020-05-29… 2020-05-… 2020-05-24     9    300
#>  7 DOH00000… 310018 RICO'… 2020-06-05… 2020-06-… 2020-05-31     9    300
#>  8 DOH00000… 311669 RICO'… 2020-06-12… 2020-06-… 2020-06-07     9    300
#>  9 DOH00000… 313263 RICO'… 2020-06-19… 2020-06-… 2020-06-14     9    300
#> 10 DOH00000… 315173 RICO'… 2020-06-26… 2020-06-… 2020-06-21     9    300
#> # … with 55,834 more rows, and 18 more variables: head_cover <dbl>,
#> #   goggles <dbl>, coverall <dbl>, shoe_cover <dbl>, face_shield <dbl>,
#> #   surgmask <dbl>, n95mask <dbl>, ops <dbl>, nps <dbl>, vtm <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest testing aggregates information
datadrop_get_tests(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 07 Testing Aggregates.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec540103990ff

#> # A tibble: 19,572 x 18
#>    facility_name report_date avg_turnaround_… daily_output_sa… daily_output_un…
#>    <chr>         <chr>                  <dbl>            <dbl>            <dbl>
#>  1 AL Molecular… 2020-08-30                24                0                0
#>  2 AL Molecular… 2020-08-31                24                0                0
#>  3 AL Molecular… 2020-09-01                NA               NA               NA
#>  4 AL Molecular… 2020-09-02                NA               NA               NA
#>  5 AL Molecular… 2020-09-03                NA               NA               NA
#>  6 AL Molecular… 2020-09-04                NA               NA               NA
#>  7 AL Molecular… 2020-09-05                NA               NA               NA
#>  8 AL Molecular… 2020-09-06                NA               NA               NA
#>  9 AL Molecular… 2020-09-07                 8                1                1
#> 10 AL Molecular… 2020-09-08                 8                1                1
#> # … with 19,562 more rows, and 13 more variables:
#> #   daily_output_positive_individuals <dbl>,
#> #   daily_output_negative_individuals <dbl>, daily_output_equivocal <dbl>,
#> #   daily_output_invalid <dbl>, remaining_available_tests <dbl>,
#> #   backlogs <dbl>, cumulative_samples_tested <int>,
#> #   cumulative_unique_individuals <int>, cumulative_positive_individuals <int>,
#> #   cumulative_negative_individuals <int>, pct_positive_cumulative <dbl>,
#> #   pct_negative_cumulative <dbl>, validation_status <chr>

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 08 Quarantine Facility Data - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec540335e5dab

#> # A tibble: 1,781,154 x 22
#>    updateddate addeddate reportdate other_adm doctor_adm nurse_adm qother
#>    <chr>       <chr>     <chr>          <dbl>      <dbl>     <dbl>  <dbl>
#>  1 2020-06-24… 2020-06-… 2020-06-2…         0          0         0      0
#>  2 2020-06-25… 2020-06-… 2020-06-2…         0          0         0      0
#>  3 2020-06-26… 2020-06-… 2020-06-2…         0          0         0      0
#>  4 2020-06-27… 2020-06-… 2020-06-2…         0          0         0      0
#>  5 2020-06-28… 2020-06-… 2020-06-2…         0          0         0      0
#>  6 2020-06-29… 2020-06-… 2020-06-2…         0          0         0      0
#>  7 2020-06-30… 2020-06-… 2020-06-3…         0          0         0      0
#>  8 2020-07-01… 2020-07-… 2020-07-0…         0          0         0      0
#>  9 2020-07-02… 2020-07-… 2020-07-0…         0          0         0      0
#> 10 2020-07-03… 2020-07-… 2020-07-0…         0          0         0      0
#> # … with 1,781,144 more rows, and 15 more variables: qdoctor <dbl>,
#> #   qnurse <dbl>, isolbed_o <dbl>, isolbed_v <dbl>, beds_ward_o <dbl>,
#> #   beds_ward_v <dbl>, facilityname <chr>, region <chr>, province <chr>,
#> #   city_mun <chr>, bgy <chr>, region_psgc <chr>, province_psgc <chr>,
#> #   city_mun_psgc <chr>, bgy_psgc <chr>

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201118 - 09 Quarantine Facility Data - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpJ7Na9I/filec54078579fe5

#> # A tibble: 1,384 x 25
#>    facilityname updateddate addeddate reportdate coverall surgmask shoe_cover
#>    <chr>        <chr>       <chr>     <chr>         <dbl>    <dbl>      <dbl>
#>  1 BAYACABAC B… ""          ""        2020-06-0…       NA       NA         NA
#>  2 DEPARTMENT … "2020-06-0… "2020-06… 2020-05-3…      185     3850         34
#>  3 DEPARTMENT … ""          ""        2020-06-0…       NA       NA         NA
#>  4 DINGALAN CO… "2020-05-2… "2020-05… 2020-05-2…        0        0          0
#>  5 DINGALAN CO… "2020-06-0… "2020-06… 2020-05-3…        0        0          0
#>  6 DINGALAN CO… ""          ""        2020-06-0…       NA       NA         NA
#>  7 MALAGANA - … "2020-06-0… "2020-06… 2020-05-3…        1       50          1
#>  8 MALAGANA - … ""          ""        2020-06-0…       NA       NA         NA
#>  9 TAMBOBOAN -… "2020-06-0… "2020-06… 2020-05-3…        1       50          1
#> 10 TAMBOBOAN -… ""          ""        2020-06-0…       NA       NA         NA
#> # … with 1,374 more rows, and 18 more variables: face_shield <dbl>,
#> #   head_cover <dbl>, n95mask <dbl>, goggles <dbl>, gloves <dbl>, gown <dbl>,
#> #   vtm <dbl>, nps <dbl>, ops <dbl>, weekly_reportdate <chr>, region <chr>,
#> #   province <chr>, city_mun <chr>, bgy <chr>, region_psgc <chr>,
#> #   province_psgc <chr>, city_mun_psgc <chr>, bgy_psgc <chr>

#datadrop_get_collectV3(tbl = x, path = tempfile())
#datadrop_get_collectV4(tbl = x, path = tempfile())
#datadrop_get_tracker(tbl = x, path = tempfile())