Retrieve datasets from specified DoH Data Drop folders

A wrapper to googledrive functions to retrieve datasets from the DoH Data Drop folders

datadrop_changelog(version = c("latest", "archive"), .date = NULL)

datadrop_sheets(version = c("latest", "archive"), .date = NULL)

datadrop_fields(version = c("latest", "archive"), .date = NULL)

datadrop_cases(version = c("latest", "archive"), .date = NULL)

datadrop_tests(version = c("latest", "archive"), .date = NULL)

datadrop_collect_daily(version = c("latest", "archive"), .date = NULL)

datadrop_collect_weekly(version = c("latest", "archive"), .date = NULL)

datadrop_quarantine_daily(version = c("latest", "archive"), .date = NULL)

datadrop_quarantine_weekly(version = c("latest", "archive"), .date = NULL)

Arguments

version
A character value specifying whether to get the latest available dataset (latest) or to get archive data (archive). Default to latest.

.date
A character value for date in YYYY-MM-DD format. This is the date for the archive DoH Data Drop for which an ID is to be returned. Should be specified when using version is set to archive otherwise ignored.

Value

A tibble of any of the following datasets:

Metadata - Sheets;

Metadata - Fields;

Case Information;

DOH Data Collect - Daily Report;

DOH Data Collect - Weekly Report;

Testing Aggregates;

Quarantine Facility Data - Daily Report;

Quarantine Facility Data - Weekly Report;

DOH Data Collect v3 - Baseline;

DOH Data Collect v4 - Baseline; and,

DDC TTMF Tracker v1. For Changelog, a named list of two tibbles - List of Changes and Most Common Changes.

Examples

datadrop_changelog()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> $`List of Changes`
#> # A tibble: 6,671 x 4
#>    CaseCode Column            `Old Value`  `New Value` 
#>    <chr>    <chr>             <chr>        <chr>       
#>  1 C100029  HealthStatus      MILD         ASYMPTOMATIC
#>  2 C100029  Admitted          YES          NO          
#>  3 C100029  DateOnset         44128.0      44127.0     
#>  4 C100324  DateSpecimen      44085.0      44115.0     
#>  5 C100324  DateResultRelease 44085.0      44115.0     
#>  6 C100338  Age               1.0          5.0         
#>  7 C100338  AgeGroup          0 to 4       5 to 9      
#>  8 C100725  RemovalType       NA           RECOVERED   
#>  9 C100725  HealthStatus      ASYMPTOMATIC RECOVERED   
#> 10 C100725  DateRecover       NA           44145.0     
#> # … with 6,661 more rows
#> 
#> $`Most Common Changes`
#> # A tibble: 1,722 x 4
#>    Column       `Old Value`  `New Value`  `Records Changed`
#>    <chr>        <chr>        <chr>                    <dbl>
#>  1 Quarantined  NO           YES                        548
#>  2 RemovalType  NA           RECOVERED                  459
#>  3 HealthStatus MILD         ASYMPTOMATIC               285
#>  4 HealthStatus ASYMPTOMATIC RECOVERED                  224
#>  5 HealthStatus MILD         RECOVERED                  217
#>  6 DateRecover  NA           44145.0                    153
#>  7 Admitted     NO           YES                        104
#>  8 DateRecover  NA           44144.0                     99
#>  9 DateRecover  NA           44146.0                     81
#> 10 DateRecover  NA           44127.0                     77
#> # … with 1,712 more rows
#> 

datadrop_sheets()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 10 x 5
#>    Sheet     Description          Source       Contact.Informa… Data.Update.Sch…
#>    <chr>     <chr>                <chr>        <chr>            <chr>           
#>  1 Case Inf… "List of confirmed … DOH - Epide… covidtracker@do… Daily at 4PM    
#>  2 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  3 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  4 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  5 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  6 DDC TTMF… "Self-reported staf… Collated by… covidtracker@do… Daily at 4PM    
#>  7 DOH Data… "Self-reported coun… Collated by… covidtracker@do… Daily at 4PM    
#>  8 DOH Data… "Self-reported IPC … Collated by… covidtracker@do… Daily at 4PM    
#>  9 Testing … ""                   DOH          covidtracker@do… Daily at 4PM    
#> 10 Privacy … ""                   -            -                -               

datadrop_fields()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 199 x 7
#>    Sheet  Field  Description      Type  Question.Format Possible.Respon… Purpose
#>    <chr>  <chr>  <chr>            <chr> <chr>           <chr>            <chr>  
#>  1 Case … CaseC… Random code ass… Text  ""              ""               ""     
#>  2 Case … Age    Age              Numb… ""              ""               ""     
#>  3 Case … AgeGr… Five-year age g… Text  ""              ""               ""     
#>  4 Case … Sex    Sex              Text  ""              ""               ""     
#>  5 Case … DateS… Date when speci… Date  ""              ""               ""     
#>  6 Case … DateR… Date of release… Date  ""              ""               ""     
#>  7 Case … DateR… Date publicly a… Date  ""              ""               ""     
#>  8 Case … DateD… Date died        Date  ""              ""               ""     
#>  9 Case … DateR… Date recovered   Date  ""              ""               ""     
#> 10 Case … Remov… Type of removal… Text  ""              ""               ""     
#> # … with 189 more rows

datadrop_cases()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 404,713 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C320893     11 10 to 14 FEMA… 2020-04-21   2020-05-02       2020-05-31 
#>  2 C334572     57 55 to 59 MALE  2020-07-13   2020-07-16       2020-08-08 
#>  3 C648959     58 55 to 59 MALE  2020-06-09   2020-06-12       2020-06-14 
#>  4 C827571     58 55 to 59 FEMA… 2020-07-12   2020-07-13       2020-08-08 
#>  5 C899164     30 30 to 34 MALE  2020-05-05   2020-05-15       2020-05-15 
#>  6 C172636     27 25 to 29 MALE  2020-08-06   2020-08-09       2020-08-12 
#>  7 C560298     33 30 to 34 MALE  2020-05-23   2020-05-24       2020-05-28 
#>  8 C493432     25 25 to 29 FEMA… 2020-06-25   2020-06-27       2020-06-30 
#>  9 C855974     25 25 to 29 FEMA… 2020-07-10   2020-07-11       2020-07-17 
#> 10 C264567     29 25 to 29 FEMA… 2020-05-09   2020-06-09       2020-07-26 
#> # … with 404,703 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

datadrop_collect_daily()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 169,677 x 54
#>    hfhudcode    id cfname updateddate addeddate reportdate icu_v icu_o isolbed_v
#>    <chr>     <dbl> <chr>  <chr>       <chr>     <chr>      <dbl> <dbl>     <dbl>
#>  1 DOH00000…   100 "SORS… 2020-07-02… 2020-07-… 2020-07-0…     0     2         1
#>  2 DOH00000…  1005 "SORS… 2020-07-03… 2020-07-… 2020-07-0…     0     2         1
#>  3 DOH00000…  2624 "SORS… 2020-07-04… 2020-07-… 2020-07-0…     0     2         1
#>  4 DOH00000…  4273 "SORS… 2020-07-05… 2020-07-… 2020-07-0…     0     2         1
#>  5 DOH00000…  5833 "SORS… 2020-07-06… 2020-07-… 2020-07-0…     1     1         2
#>  6 DOH00000…  7631 "SORS… 2020-07-07… 2020-07-… 2020-07-0…     1     1         2
#>  7 DOH00000…  9352 "SORS… 2020-07-08… 2020-07-… 2020-07-0…     1     1         2
#>  8 DOH00000… 11175 "SORS… 2020-07-09… 2020-07-… 2020-07-0…     1     1         3
#>  9 DOH00000… 12929 "SORS… 2020-07-10… 2020-07-… 2020-07-1…     2     0         3
#> 10 DOH00000… 14791 "SORS… 2020-07-11… 2020-07-… 2020-07-1…     2     0         3
#> # … with 169,667 more rows, and 45 more variables: isolbed_o <dbl>,
#> #   beds_ward_v <dbl>, beds_ward_o <dbl>, mechvent_v <dbl>, mechvent_o <dbl>,
#> #   icu_v_nc <dbl>, icu_o_nc <dbl>, nonicu_v_nc <dbl>, nonicu_o_nc <dbl>,
#> #   mechvent_v_nc <dbl>, mechvent_o_nc <dbl>, qnurse <dbl>, qdoctor <dbl>,
#> #   qother <dbl>, nurse_adm <dbl>, doctor_adm <dbl>, other_adm <dbl>,
#> #   susp_asym <dbl>, susp_mild <dbl>, susp_severe <dbl>, susp_crit <dbl>,
#> #   susp_died <dbl>, prob_asym <dbl>, prob_mild <dbl>, prob_severe <dbl>,
#> #   prob_crit <dbl>, prob_died <dbl>, conf_asym <dbl>, conf_mild <dbl>,
#> #   conf_severe <dbl>, conf_crit <dbl>, conf_died <dbl>, tpatient_adm <dbl>,
#> #   tpatient_er <dbl>, tpatient_icu <dbl>, trans_ttmf <dbl>, discharged <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

datadrop_collect_weekly()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 53,945 x 26
#>    hfhudcode     id cfname updateddate addeddate reportdate  gown gloves
#>    <chr>      <dbl> <chr>  <chr>       <chr>     <chr>      <dbl>  <dbl>
#>  1 DOH00000…  31743 OUR L… 2020-04-24… 2020-04-… 2020-04-19    35   3300
#>  2 DOH00000…  32651 OUR L… 2020-05-01… 2020-05-… 2020-04-26    45   6800
#>  3 DOH00000…  33683 OUR L… 2020-05-08… 2020-05-… 2020-05-03    45   5500
#>  4 DOH00000…  36018 OUR L… 2020-05-15… 2020-05-… 2020-05-10    45  14000
#>  5 DOH00000…  37253 OUR L… 2020-05-22… 2020-05-… 2020-05-17    95  15200
#>  6 DOH00000…  38716 OUR L… 2020-05-29… 2020-05-… 2020-05-24   100  10130
#>  7 DOH00000… 310282 OUR L… 2020-06-05… 2020-06-… 2020-05-31   100   9100
#>  8 DOH00000… 313042 OUR L… 2020-06-12… 2020-06-… 2020-06-07   100   6400
#>  9 DOH00000… 313588 OUR L… 2020-06-19… 2020-06-… 2020-06-14   100   5900
#> 10 DOH00000… 315233 OUR L… 2020-06-26… 2020-06-… 2020-06-21   100   4500
#> # … with 53,935 more rows, and 18 more variables: head_cover <dbl>,
#> #   goggles <dbl>, coverall <dbl>, shoe_cover <dbl>, face_shield <dbl>,
#> #   surgmask <dbl>, n95mask <dbl>, ops <dbl>, nps <dbl>, vtm <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

datadrop_tests()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 18,819 x 18
#>    facility_name report_date avg_turnaround_… daily_output_sa… daily_output_un…
#>    <chr>         <chr>                  <dbl>            <dbl>            <dbl>
#>  1 AL Molecular… 2020-08-30                24                0                0
#>  2 AL Molecular… 2020-08-31                24                0                0
#>  3 AL Molecular… 2020-09-01                NA               NA               NA
#>  4 AL Molecular… 2020-09-02                NA               NA               NA
#>  5 AL Molecular… 2020-09-03                NA               NA               NA
#>  6 AL Molecular… 2020-09-04                NA               NA               NA
#>  7 AL Molecular… 2020-09-05                NA               NA               NA
#>  8 AL Molecular… 2020-09-06                NA               NA               NA
#>  9 AL Molecular… 2020-09-07                 8                1                1
#> 10 AL Molecular… 2020-09-08                 8                1                1
#> # … with 18,809 more rows, and 13 more variables:
#> #   daily_output_positive_individuals <dbl>,
#> #   daily_output_negative_individuals <dbl>, daily_output_equivocal <dbl>,
#> #   daily_output_invalid <dbl>, remaining_available_tests <dbl>,
#> #   backlogs <dbl>, cumulative_samples_tested <int>,
#> #   cumulative_unique_individuals <int>, cumulative_positive_individuals <int>,
#> #   cumulative_negative_individuals <int>, pct_positive_cumulative <dbl>,
#> #   pct_negative_cumulative <dbl>, validation_status <chr>

datadrop_quarantine_daily()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 1,709,046 x 22
#>    updateddate addeddate reportdate other_adm doctor_adm nurse_adm qother
#>    <chr>       <chr>     <chr>          <dbl>      <dbl>     <dbl>  <dbl>
#>  1 2020-05-18… 2020-05-… 2020-05-1…         0          0         0      0
#>  2 2020-05-19… 2020-05-… 2020-05-1…         0          0         0      0
#>  3 2020-05-20… 2020-05-… 2020-05-2…         0          0         0      0
#>  4 2020-05-21… 2020-05-… 2020-05-2…         0          0         0      0
#>  5 2020-05-22… 2020-05-… 2020-05-2…         0          0         0      0
#>  6 2020-05-23… 2020-05-… 2020-05-2…         0          0         0      0
#>  7 2020-05-24… 2020-05-… 2020-05-2…         0          0         0      0
#>  8 2020-05-25… 2020-05-… 2020-05-2…         0          0         0      0
#>  9 2020-05-26… 2020-05-… 2020-05-2…         0          0         0      0
#> 10 2020-05-27… 2020-05-… 2020-05-2…         0          0         0      0
#> # … with 1,709,036 more rows, and 15 more variables: qdoctor <dbl>,
#> #   qnurse <dbl>, isolbed_o <dbl>, isolbed_v <dbl>, beds_ward_o <dbl>,
#> #   beds_ward_v <dbl>, facilityname <chr>, region <chr>, province <chr>,
#> #   city_mun <chr>, bgy <chr>, region_psgc <chr>, province_psgc <chr>,
#> #   city_mun_psgc <chr>, bgy_psgc <chr>

datadrop_quarantine_weekly()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

#> # A tibble: 1,387 x 25
#>    facilityname updateddate addeddate reportdate coverall surgmask shoe_cover
#>    <chr>        <chr>       <chr>     <chr>         <dbl>    <dbl>      <dbl>
#>  1 BAUNOS ISOL… ""          ""        2020-06-0…       NA       NA         NA
#>  2 KIAGOT BARA… ""          ""        2020-06-0…       NA       NA         NA
#>  3 SIBUTAD EVA… ""          ""        2020-06-0…       NA       NA         NA
#>  4 TAMBONGON Q… ""          ""        2020-06-0…       NA       NA         NA
#>  5 TUBAY BAHAY… "2020-06-0… "2020-06… 2020-05-3…        0      200          0
#>  6 TUBAY BAHAY… ""          ""        2020-06-0…       NA       NA         NA
#>  7 BADIANG DAY… ""          ""        2020-06-0…       NA       NA         NA
#>  8 FATIMA BARA… ""          ""        2020-06-0…       NA       NA         NA
#>  9 HALF WAY HO… "2020-05-2… "2020-05… 2020-05-2…       87     2000         87
#> 10 HALF WAY HO… "2020-05-3… "2020-05… 2020-05-3…       87     1900         87
#> # … with 1,377 more rows, and 18 more variables: face_shield <dbl>,
#> #   head_cover <dbl>, n95mask <dbl>, goggles <dbl>, gloves <dbl>, gown <dbl>,
#> #   vtm <dbl>, nps <dbl>, ops <dbl>, weekly_reportdate <chr>, region <chr>,
#> #   province <chr>, city_mun <chr>, bgy <chr>, region_psgc <chr>,
#> #   province_psgc <chr>, city_mun_psgc <chr>, bgy_psgc <chr>

#datadrop_collect_v3()
#datadrop_collect_v4()
#datadrop_tracker()