Retrieve datasets from specified DoH Data Drop folders

A wrapper to googledrive functions to retrieve datasets from the DoH Data Drop folders.

datadrop_get(
  tbl,
  fn,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_changelog(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

tbl
A tibble output produced by datadrop_ls() that lists the files within a particular DoH Data Drop Google Drive folder

fn
A character string composed of a word or words that can be used to match to the name of a file within a particular DoH Data Drop Google Drive folder listed in tbl.

path
A character value for path for output file. If NULL, the default file name used in Google Drive is used and the default location is the working directory.

keep
Logical. Should file be saved locally? Default to FALSE. If TRUE, file is kept in the location specified in path. If path is NULL, the file is kept in the working directory using the default file name used in Google Drive.

overwrite
Logical. If path already exists, should it be overwritten? Default to FALSE.

verbose
Logical. Should operation progress messages be shown? Default to TRUE.

Value

A tibble of any of the following datasets:

Metadata - Sheets;

Metadata - Fields;

Case Information;

DOH Data Collect - Daily Report;

DOH Data Collect - Weekly Report;

Testing Aggregates;

Quarantine Facility Data - Daily Report;

Quarantine Facility Data - Weekly Report;

DOH Data Collect v3 - Baseline;

DOH Data Collect v4 - Baseline; and,

DDC TTMF Tracker v1. For Changelog, a named list of two tibbles - List of Changes and Most Common Changes. If keep is TRUE, a copy of the specified DoH Data Drop Google Drive file is saved in the location specified by path.

Examples

## Get tbl for files in latest Data Drop
library(magrittr)
x <- datadrop_id() %>% datadrop_ls()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-20.

#> File downloaded:
#>   * 01 READ ME FIRST (11_20).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f3b6ea779

## Retrieve case information data
datadrop_get(tbl = x, fn = "Case Information", path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f5db5176a

#> # A tibble: 415,067 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C107690     23 20 to 24 MALE  "2020-07-17" 2020-07-24       2020-07-28 
#>  2 C198106     40 40 to 44 MALE  "2020-07-08" 2020-07-11       2020-07-23 
#>  3 C223825     25 25 to 29 MALE  "2020-07-22" 2020-07-27       2020-08-01 
#>  4 C145699     49 45 to 49 MALE  ""           2020-08-04       2020-08-08 
#>  5 C991185     62 60 to 64 FEMA… "2020-07-21" 2020-07-26       2020-07-31 
#>  6 C502968     42 40 to 44 MALE  "2020-08-07" 2020-08-07       2020-08-10 
#>  7 C970007     47 45 to 49 MALE  "2020-05-06" 2020-05-08       2020-05-29 
#>  8 C220019     32 30 to 34 MALE  "2020-03-10" 2020-07-12       2020-07-18 
#>  9 C690794     35 35 to 39 MALE  "2020-07-14" 2020-07-16       2020-07-19 
#> 10 C475319     27 25 to 29 MALE  "2020-07-16" 2020-07-25       2020-08-01 
#> # … with 415,057 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest changelog information
datadrop_get_changelog(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH Data Drop 20201120 - Changelog.xlsx
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f1c146036

#> $`List of Changes`
#> # A tibble: 5,591 x 4
#>    CaseCode Column            `Old Value`  `New Value`
#>    <chr>    <chr>             <chr>        <chr>      
#>  1 C100065  DateRecover       NA           44153.0    
#>  2 C100065  RemovalType       NA           RECOVERED  
#>  3 C100065  HealthStatus      ASYMPTOMATIC RECOVERED  
#>  4 C100119  DateSpecimen      43898.0      NA         
#>  5 C100350  Quarantined       NO           YES        
#>  6 C100760  Quarantined       NO           YES        
#>  7 C100760  DateOnset         NA           44129.0    
#>  8 C100772  DateRecover       44136.0      44133.0    
#>  9 C100815  DateSpecimen      44135.0      NA         
#> 10 C100815  DateResultRelease 44135.0      NA         
#> # … with 5,581 more rows
#> 
#> $`Most Common Changes`
#> # A tibble: 1,536 x 4
#>    Column            `Old Value`  `New Value`  `Records Changed`
#>    <chr>             <chr>        <chr>                    <dbl>
#>  1 Quarantined       NO           YES                        636
#>  2 RemovalType       NA           RECOVERED                  283
#>  3 HealthStatus      MILD         ASYMPTOMATIC               223
#>  4 HealthStatus      MILD         RECOVERED                  137
#>  5 HealthStatus      ASYMPTOMATIC RECOVERED                  130
#>  6 DateRecover       NA           44152.0                    125
#>  7 DateRecover       NA           44133.0                    108
#>  8 Admitted          NO           YES                         70
#>  9 DateRecover       NA           44134.0                     63
#> 10 DateResultRelease 44141.0      NA                          40
#> # … with 1,526 more rows
#> 

## Retrieve latest metadata - sheets information
datadrop_get_sheets(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 02 Metadata - Sheets.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f7db887a0

#> # A tibble: 10 x 5
#>    Sheet     Description          Source       Contact.Informa… Data.Update.Sch…
#>    <chr>     <chr>                <chr>        <chr>            <chr>           
#>  1 Case Inf… "List of confirmed … DOH - Epide… covidtracker@do… Daily at 4PM    
#>  2 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  3 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  4 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  5 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  6 DDC TTMF… "Self-reported staf… Collated by… covidtracker@do… Daily at 4PM    
#>  7 DOH Data… "Self-reported coun… Collated by… covidtracker@do… Daily at 4PM    
#>  8 DOH Data… "Self-reported IPC … Collated by… covidtracker@do… Daily at 4PM    
#>  9 Testing … ""                   DOH          covidtracker@do… Daily at 4PM    
#> 10 Privacy … ""                   -            -                -               

## Retrieve latest metadata - fields information
datadrop_get_fields(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 03 Metadata - Fields.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f5dd059db

#> # A tibble: 199 x 7
#>    Sheet  Field  Description      Type  Question.Format Possible.Respon… Purpose
#>    <chr>  <chr>  <chr>            <chr> <chr>           <chr>            <chr>  
#>  1 Case … CaseC… Random code ass… Text  ""              ""               ""     
#>  2 Case … Age    Age              Numb… ""              ""               ""     
#>  3 Case … AgeGr… Five-year age g… Text  ""              ""               ""     
#>  4 Case … Sex    Sex              Text  ""              ""               ""     
#>  5 Case … DateS… Date when speci… Date  ""              ""               ""     
#>  6 Case … DateR… Date of release… Date  ""              ""               ""     
#>  7 Case … DateR… Date publicly a… Date  ""              ""               ""     
#>  8 Case … DateD… Date died        Date  ""              ""               ""     
#>  9 Case … DateR… Date recovered   Date  ""              ""               ""     
#> 10 Case … Remov… Type of removal… Text  ""              ""               ""     
#> # … with 189 more rows

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f19bb68fb

#> # A tibble: 415,067 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C107690     23 20 to 24 MALE  "2020-07-17" 2020-07-24       2020-07-28 
#>  2 C198106     40 40 to 44 MALE  "2020-07-08" 2020-07-11       2020-07-23 
#>  3 C223825     25 25 to 29 MALE  "2020-07-22" 2020-07-27       2020-08-01 
#>  4 C145699     49 45 to 49 MALE  ""           2020-08-04       2020-08-08 
#>  5 C991185     62 60 to 64 FEMA… "2020-07-21" 2020-07-26       2020-07-31 
#>  6 C502968     42 40 to 44 MALE  "2020-08-07" 2020-08-07       2020-08-10 
#>  7 C970007     47 45 to 49 MALE  "2020-05-06" 2020-05-08       2020-05-29 
#>  8 C220019     32 30 to 34 MALE  "2020-03-10" 2020-07-12       2020-07-18 
#>  9 C690794     35 35 to 39 MALE  "2020-07-14" 2020-07-16       2020-07-19 
#> 10 C475319     27 25 to 29 MALE  "2020-07-16" 2020-07-25       2020-08-01 
#> # … with 415,057 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 05 DOH Data Collect - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f5ee943ef

#> # A tibble: 178,644 x 54
#>    hfhudcode    id cfname updateddate addeddate reportdate icu_v icu_o isolbed_v
#>    <chr>     <dbl> <chr>  <chr>       <chr>     <chr>      <dbl> <dbl>     <dbl>
#>  1 DOH00000…   432 Minda… 2020-07-02… 2020-07-… 2020-07-0…     0     0         6
#>  2 DOH00000…  1531 Minda… 2020-07-03… 2020-07-… 2020-07-0…     0     0         6
#>  3 DOH00000…  3135 Minda… 2020-07-04… 2020-07-… 2020-07-0…     0     0         6
#>  4 DOH00000…  4704 Minda… 2020-07-05… 2020-07-… 2020-07-0…     0     0         4
#>  5 DOH00000…  6569 Minda… 2020-07-06… 2020-07-… 2020-07-0…     0     0         3
#>  6 DOH00000…  8680 Minda… 2020-07-07… 2020-07-… 2020-07-0…     0     0         3
#>  7 DOH00000… 10314 Minda… 2020-07-08… 2020-07-… 2020-07-0…     0     0         5
#>  8 DOH00000… 11819 Minda… 2020-07-09… 2020-07-… 2020-07-0…     0     0         3
#>  9 DOH00000… 13605 Minda… 2020-07-10… 2020-07-… 2020-07-1…     0     0         0
#> 10 DOH00000… 15562 Minda… 2020-07-11… 2020-07-… 2020-07-1…     0     0         0
#> # … with 178,634 more rows, and 45 more variables: isolbed_o <dbl>,
#> #   beds_ward_v <dbl>, beds_ward_o <dbl>, mechvent_v <dbl>, mechvent_o <dbl>,
#> #   icu_v_nc <dbl>, icu_o_nc <dbl>, nonicu_v_nc <dbl>, nonicu_o_nc <dbl>,
#> #   mechvent_v_nc <dbl>, mechvent_o_nc <dbl>, qnurse <dbl>, qdoctor <dbl>,
#> #   qother <dbl>, nurse_adm <dbl>, doctor_adm <dbl>, other_adm <dbl>,
#> #   susp_asym <dbl>, susp_mild <dbl>, susp_severe <dbl>, susp_crit <dbl>,
#> #   susp_died <dbl>, prob_asym <dbl>, prob_mild <dbl>, prob_severe <dbl>,
#> #   prob_crit <dbl>, prob_died <dbl>, conf_asym <dbl>, conf_mild <dbl>,
#> #   conf_severe <dbl>, conf_crit <dbl>, conf_died <dbl>, tpatient_adm <dbl>,
#> #   tpatient_er <dbl>, tpatient_icu <dbl>, trans_ttmf <dbl>, discharged <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 06 DOH Data Collect - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f246b3097

#> # A tibble: 55,844 x 26
#>    hfhudcode     id cfname updateddate addeddate reportdate  gown gloves
#>    <chr>      <dbl> <chr>  <chr>       <chr>     <chr>      <dbl>  <dbl>
#>  1 DOH00000…  32104 UMALI… 2020-05-01… 2020-05-… 2020-04-26     0      0
#>  2 DOH00000…  34153 UMALI… 2020-05-08… 2020-05-… 2020-05-03     5      0
#>  3 DOH00000…  35157 UMALI… 2020-05-15… 2020-05-… 2020-05-10     5      0
#>  4 DOH00000…  36819 UMALI… 2020-05-22… 2020-05-… 2020-05-17     5      0
#>  5 DOH00000…  38484 UMALI… 2020-05-29… 2020-05-… 2020-05-24     5      0
#>  6 DOH00000… 310348 UMALI… 2020-06-05… 2020-06-… 2020-05-31    18      0
#>  7 DOH00000… 311774 UMALI… 2020-06-12… 2020-06-… 2020-06-07    18      0
#>  8 DOH00000… 313428 UMALI… 2020-06-19… 2020-06-… 2020-06-14    18      0
#>  9 DOH00000… 314970 UMALI… 2020-06-26… 2020-06-… 2020-06-21    18      0
#> 10 DOH00000…  41044 UMALI… 2020-07-03… 2020-07-… 2020-06-28    18      0
#> # … with 55,834 more rows, and 18 more variables: head_cover <dbl>,
#> #   goggles <dbl>, coverall <dbl>, shoe_cover <dbl>, face_shield <dbl>,
#> #   surgmask <dbl>, n95mask <dbl>, ops <dbl>, nps <dbl>, vtm <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest testing aggregates information
datadrop_get_tests(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 07 Testing Aggregates.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f79431c2e

#> # A tibble: 19,902 x 18
#>    facility_name report_date avg_turnaround_… daily_output_sa… daily_output_un…
#>    <chr>         <chr>                  <dbl>            <dbl>            <dbl>
#>  1 AL Molecular… 2020-08-30                24                0                0
#>  2 AL Molecular… 2020-08-31                24                0                0
#>  3 AL Molecular… 2020-09-01                NA               NA               NA
#>  4 AL Molecular… 2020-09-02                NA               NA               NA
#>  5 AL Molecular… 2020-09-03                NA               NA               NA
#>  6 AL Molecular… 2020-09-04                NA               NA               NA
#>  7 AL Molecular… 2020-09-05                NA               NA               NA
#>  8 AL Molecular… 2020-09-06                NA               NA               NA
#>  9 AL Molecular… 2020-09-07                 8                1                1
#> 10 AL Molecular… 2020-09-08                 8                1                1
#> # … with 19,892 more rows, and 13 more variables:
#> #   daily_output_positive_individuals <dbl>,
#> #   daily_output_negative_individuals <dbl>, daily_output_equivocal <dbl>,
#> #   daily_output_invalid <dbl>, remaining_available_tests <dbl>,
#> #   backlogs <dbl>, cumulative_samples_tested <int>,
#> #   cumulative_unique_individuals <int>, cumulative_positive_individuals <int>,
#> #   cumulative_negative_individuals <int>, pct_positive_cumulative <dbl>,
#> #   pct_negative_cumulative <dbl>, validation_status <chr>

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 08 Quarantine Facility Data - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f24ef4e34

#> # A tibble: 1,811,628 x 22
#>    updateddate addeddate reportdate other_adm doctor_adm nurse_adm qother
#>    <chr>       <chr>     <chr>          <dbl>      <dbl>     <dbl>  <dbl>
#>  1 2020-08-17… 2020-08-… 2020-08-1…         0          0         0      0
#>  2 2020-08-18… 2020-08-… 2020-08-1…         0          0         0      0
#>  3 2020-08-19… 2020-08-… 2020-08-1…         0          0         0      0
#>  4 2020-08-20… 2020-08-… 2020-08-2…         0          0         0      0
#>  5 2020-08-21… 2020-08-… 2020-08-2…         0          0         0      0
#>  6 2020-08-22… 2020-08-… 2020-08-2…         0          0         0      0
#>  7 2020-08-23… 2020-08-… 2020-08-2…         0          0         0      0
#>  8 2020-08-24… 2020-08-… 2020-08-2…         0          0         0      0
#>  9 2020-08-25… 2020-08-… 2020-08-2…         0          0         0      0
#> 10 2020-08-26… 2020-08-… 2020-08-2…         0          0         0      0
#> # … with 1,811,618 more rows, and 15 more variables: qdoctor <dbl>,
#> #   qnurse <dbl>, isolbed_o <dbl>, isolbed_v <dbl>, beds_ward_o <dbl>,
#> #   beds_ward_v <dbl>, facilityname <chr>, region <chr>, province <chr>,
#> #   city_mun <chr>, bgy <chr>, region_psgc <chr>, province_psgc <chr>,
#> #   city_mun_psgc <chr>, bgy_psgc <chr>

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201120 - 09 Quarantine Facility Data - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmphXmxsg/filebb2f5af74add

#> # A tibble: 1,382 x 25
#>    facilityname updateddate addeddate reportdate coverall surgmask shoe_cover
#>    <chr>        <chr>       <chr>     <chr>         <dbl>    <dbl>      <dbl>
#>  1 SAN JUAN DA… ""          ""        2020-06-0…       NA       NA         NA
#>  2 NINOY AQUIN… "2020-06-0… "2020-06… 2020-05-3…    10261    38760       8162
#>  3 NINOY AQUIN… ""          ""        2020-06-0…       NA       NA         NA
#>  4 NIPA ELEMEN… ""          ""        2020-06-0…       NA       NA         NA
#>  5 BARANGAY CO… ""          ""        2020-05-2…       NA       NA         NA
#>  6 BARANGAY CO… "2020-06-0… "2020-06… 2020-05-3…        0        0          0
#>  7 BARANGAY CO… ""          ""        2020-06-0…       NA       NA         NA
#>  8 ESPERANZA D… ""          ""        2020-06-0…       NA       NA         NA
#>  9 SOLANA SENI… ""          ""        2020-06-0…       NA       NA         NA
#> 10 CASIGURAN M… "2020-05-2… "2020-05… 2020-05-2…        2       10          5
#> # … with 1,372 more rows, and 18 more variables: face_shield <dbl>,
#> #   head_cover <dbl>, n95mask <dbl>, goggles <dbl>, gloves <dbl>, gown <dbl>,
#> #   vtm <dbl>, nps <dbl>, ops <dbl>, weekly_reportdate <chr>, region <chr>,
#> #   province <chr>, city_mun <chr>, bgy <chr>, region_psgc <chr>,
#> #   province_psgc <chr>, city_mun_psgc <chr>, bgy_psgc <chr>

#datadrop_get_collectV3(tbl = x, path = tempfile())
#datadrop_get_collectV4(tbl = x, path = tempfile())
#datadrop_get_tracker(tbl = x, path = tempfile())