Retrieve datasets from specified DoH Data Drop folders

A wrapper to googledrive functions to retrieve datasets from the DoH Data Drop folders.

datadrop_get(
  tbl,
  fn,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_changelog(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_sheets(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_fields(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cases(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_tests(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_cweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qdaily(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

datadrop_get_qweekly(
  tbl,
  path = NULL,
  keep = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

tbl
A tibble output produced by datadrop_ls() that lists the files within a particular DoH Data Drop Google Drive folder

fn
A character string composed of a word or words that can be used to match to the name of a file within a particular DoH Data Drop Google Drive folder listed in tbl.

path
A character value for path for output file. If NULL, the default file name used in Google Drive is used and the default location is the working directory.

keep
Logical. Should file be saved locally? Default to FALSE. If TRUE, file is kept in the location specified in path. If path is NULL, the file is kept in the working directory using the default file name used in Google Drive.

overwrite
Logical. If path already exists, should it be overwritten? Default to FALSE.

verbose
Logical. Should operation progress messages be shown? Default to TRUE.

Value

A tibble of any of the following datasets:

Metadata - Sheets;

Metadata - Fields;

Case Information;

DOH Data Collect - Daily Report;

DOH Data Collect - Weekly Report;

Testing Aggregates;

Quarantine Facility Data - Daily Report;

Quarantine Facility Data - Weekly Report;

DOH Data Collect v3 - Baseline;

DOH Data Collect v4 - Baseline; and,

DDC TTMF Tracker v1. For Changelog, a named list of two tibbles - List of Changes and Most Common Changes. If keep is TRUE, a copy of the specified DoH Data Drop Google Drive file is saved in the location specified by path.

Examples

## Get tbl for files in latest Data Drop
library(magrittr)
x <- datadrop_id() %>% datadrop_ls()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-19.

#> File downloaded:
#>   * 01 READ ME FIRST (11_19).pdf
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec849522d8e11

## Retrieve case information data
datadrop_get(tbl = x, fn = "Case Information", path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec8492cca283d

#> # A tibble: 413,430 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C961991     38 35 to 39 FEMA… 2020-07-20   "2020-07-24"     2020-07-28 
#>  2 C374422     51 50 to 54 FEMA… 2020-05-25   ""               2020-05-31 
#>  3 C927869     68 65 to 69 MALE  2020-07-27   "2020-07-29"     2020-08-01 
#>  4 C111115     55 55 to 59 MALE  2020-07-14   "2020-07-16"     2020-07-19 
#>  5 C711756     28 25 to 29 FEMA… 2020-06-17   ""               2020-06-23 
#>  6 C270636     28 25 to 29 MALE  2020-05-10   "2020-08-24"     2020-08-28 
#>  7 C294694     50 50 to 54 FEMA… 2020-07-22   "2020-07-25"     2020-08-02 
#>  8 C341042     52 50 to 54 MALE  2020-07-24   "2020-07-28"     2020-08-02 
#>  9 C902321     54 50 to 54 FEMA… 2020-07-01   "2020-07-04"     2020-07-06 
#> 10 C267766     52 50 to 54 FEMA… 2020-07-02   "2020-07-03"     2020-07-05 
#> # … with 413,420 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest changelog information
datadrop_get_changelog(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH Data Drop 20201119 - Changelog.xlsx
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec849c17d3c4

#> $`List of Changes`
#> # A tibble: 5,076 x 4
#>    CaseCode Column       `Old Value`               `New Value`
#>    <chr>    <chr>        <chr>                     <chr>      
#>  1 C100098  DateRecover  NA                        44150.0    
#>  2 C101823  CityMuniPSGC PH034926000               NA         
#>  3 C101823  BarangayPSGC PH034926007               NA         
#>  4 C101823  Quarantined  NO                        YES        
#>  5 C101823  BarangayRes  CAMANACSACAN              NA         
#>  6 C101823  RegionRes    Region III: Central Luzon ROF        
#>  7 C101823  CityMunRes   SAN JOSE CITY             NA         
#>  8 C101823  ProvRes      NUEVA ECIJA               NA         
#>  9 C101984  DateSpecimen NA                        44148.0    
#> 10 C101984  Admitted     NO                        YES        
#> # … with 5,066 more rows
#> 
#> $`Most Common Changes`
#> # A tibble: 1,454 x 4
#>    Column       `Old Value`  `New Value`  `Records Changed`
#>    <chr>        <chr>        <chr>                    <dbl>
#>  1 RemovalType  NA           RECOVERED                  272
#>  2 Quarantined  NO           YES                        256
#>  3 HealthStatus MILD         RECOVERED                  128
#>  4 HealthStatus ASYMPTOMATIC RECOVERED                  116
#>  5 DateRecover  NA           44135.0                    111
#>  6 DateRecover  NA           44151.0                     94
#>  7 HealthStatus MILD         ASYMPTOMATIC                93
#>  8 DateRecover  NA           44137.0                     82
#>  9 DateRecover  NA           44132.0                     79
#> 10 DateRecover  NA           44128.0                     73
#> # … with 1,444 more rows
#> 

## Retrieve latest metadata - sheets information
datadrop_get_sheets(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 02 Metadata - Sheets.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec849704fef0f

#> # A tibble: 10 x 5
#>    Sheet     Description          Source       Contact.Informa… Data.Update.Sch…
#>    <chr>     <chr>                <chr>        <chr>            <chr>           
#>  1 Case Inf… "List of confirmed … DOH - Epide… covidtracker@do… Daily at 4PM    
#>  2 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  3 DOH Data… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  4 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  5 DDC TTMF… "Self-reported inve… Collated by… covidtracker@do… Daily at 4PM    
#>  6 DDC TTMF… "Self-reported staf… Collated by… covidtracker@do… Daily at 4PM    
#>  7 DOH Data… "Self-reported coun… Collated by… covidtracker@do… Daily at 4PM    
#>  8 DOH Data… "Self-reported IPC … Collated by… covidtracker@do… Daily at 4PM    
#>  9 Testing … ""                   DOH          covidtracker@do… Daily at 4PM    
#> 10 Privacy … ""                   -            -                -               

## Retrieve latest metadata - fields information
datadrop_get_fields(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 03 Metadata - Fields.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec849fd7fb64

#> # A tibble: 199 x 7
#>    Sheet  Field  Description      Type  Question.Format Possible.Respon… Purpose
#>    <chr>  <chr>  <chr>            <chr> <chr>           <chr>            <chr>  
#>  1 Case … CaseC… Random code ass… Text  ""              ""               ""     
#>  2 Case … Age    Age              Numb… ""              ""               ""     
#>  3 Case … AgeGr… Five-year age g… Text  ""              ""               ""     
#>  4 Case … Sex    Sex              Text  ""              ""               ""     
#>  5 Case … DateS… Date when speci… Date  ""              ""               ""     
#>  6 Case … DateR… Date of release… Date  ""              ""               ""     
#>  7 Case … DateR… Date publicly a… Date  ""              ""               ""     
#>  8 Case … DateD… Date died        Date  ""              ""               ""     
#>  9 Case … DateR… Date recovered   Date  ""              ""               ""     
#> 10 Case … Remov… Type of removal… Text  ""              ""               ""     
#> # … with 189 more rows

## Retrieve latest cases information (same results as first example)
datadrop_get_cases(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 04 Case Information.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec8492cb96a5c

#> # A tibble: 413,430 x 22
#>    CaseCode   Age AgeGroup Sex   DateSpecimen DateResultRelea… DateRepConf
#>    <chr>    <dbl> <chr>    <chr> <chr>        <chr>            <chr>      
#>  1 C961991     38 35 to 39 FEMA… 2020-07-20   "2020-07-24"     2020-07-28 
#>  2 C374422     51 50 to 54 FEMA… 2020-05-25   ""               2020-05-31 
#>  3 C927869     68 65 to 69 MALE  2020-07-27   "2020-07-29"     2020-08-01 
#>  4 C111115     55 55 to 59 MALE  2020-07-14   "2020-07-16"     2020-07-19 
#>  5 C711756     28 25 to 29 FEMA… 2020-06-17   ""               2020-06-23 
#>  6 C270636     28 25 to 29 MALE  2020-05-10   "2020-08-24"     2020-08-28 
#>  7 C294694     50 50 to 54 FEMA… 2020-07-22   "2020-07-25"     2020-08-02 
#>  8 C341042     52 50 to 54 MALE  2020-07-24   "2020-07-28"     2020-08-02 
#>  9 C902321     54 50 to 54 FEMA… 2020-07-01   "2020-07-04"     2020-07-06 
#> 10 C267766     52 50 to 54 FEMA… 2020-07-02   "2020-07-03"     2020-07-05 
#> # … with 413,420 more rows, and 15 more variables: DateDied <chr>,
#> #   DateRecover <chr>, RemovalType <chr>, Admitted <chr>, RegionRes <chr>,
#> #   ProvRes <chr>, CityMunRes <chr>, CityMuniPSGC <chr>, BarangayRes <chr>,
#> #   BarangayPSGC <chr>, HealthStatus <chr>, Quarantined <chr>, DateOnset <chr>,
#> #   Pregnanttab <chr>, ValidationStatus <chr>

## Retrieve latest daily hospital beds and mechanical ventilators information
datadrop_get_cdaily(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 05 DOH Data Collect - Daily Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec84940f5d4f4

#> # A tibble: 177,363 x 54
#>    hfhudcode    id cfname updateddate addeddate reportdate icu_v icu_o isolbed_v
#>    <chr>     <dbl> <chr>  <chr>       <chr>     <chr>      <dbl> <dbl>     <dbl>
#>  1 DOH00000…  2210 SIMBU… 2020-07-03… 2020-07-… 2020-07-0…     0     0         0
#>  2 DOH00000…  3610 SIMBU… 2020-07-04… 2020-07-… 2020-07-0…     0     0         0
#>  3 DOH00000…  5057 SIMBU… 2020-07-05… 2020-07-… 2020-07-0…     0     0         0
#>  4 DOH00000…  6675 SIMBU… 2020-07-06… 2020-07-… 2020-07-0…     0     0         0
#>  5 DOH00000…  8308 SIMBU… 2020-07-07… 2020-07-… 2020-07-0…     0     0         0
#>  6 DOH00000…  9899 SIMBU… 2020-07-08… 2020-07-… 2020-07-0…     0     0         0
#>  7 DOH00000… 12003 SIMBU… 2020-07-09… 2020-07-… 2020-07-0…     0     0         0
#>  8 DOH00000… 13895 SIMBU… 2020-07-10… 2020-07-… 2020-07-1…     0     0         0
#>  9 DOH00000… 15535 SIMBU… 2020-07-11… 2020-07-… 2020-07-1…     0     0         0
#> 10 DOH00000… 17360 SIMBU… 2020-07-12… 2020-07-… 2020-07-1…     0     0         0
#> # … with 177,353 more rows, and 45 more variables: isolbed_o <dbl>,
#> #   beds_ward_v <dbl>, beds_ward_o <dbl>, mechvent_v <dbl>, mechvent_o <dbl>,
#> #   icu_v_nc <dbl>, icu_o_nc <dbl>, nonicu_v_nc <dbl>, nonicu_o_nc <dbl>,
#> #   mechvent_v_nc <dbl>, mechvent_o_nc <dbl>, qnurse <dbl>, qdoctor <dbl>,
#> #   qother <dbl>, nurse_adm <dbl>, doctor_adm <dbl>, other_adm <dbl>,
#> #   susp_asym <dbl>, susp_mild <dbl>, susp_severe <dbl>, susp_crit <dbl>,
#> #   susp_died <dbl>, prob_asym <dbl>, prob_mild <dbl>, prob_severe <dbl>,
#> #   prob_crit <dbl>, prob_died <dbl>, conf_asym <dbl>, conf_mild <dbl>,
#> #   conf_severe <dbl>, conf_crit <dbl>, conf_died <dbl>, tpatient_adm <dbl>,
#> #   tpatient_er <dbl>, tpatient_icu <dbl>, trans_ttmf <dbl>, discharged <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest weekly PPE and other related equipment information
datadrop_get_cweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 06 DOH Data Collect - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec8494b70007d

#> # A tibble: 55,844 x 26
#>    hfhudcode     id cfname updateddate addeddate reportdate  gown gloves
#>    <chr>      <dbl> <chr>  <chr>       <chr>     <chr>      <dbl>  <dbl>
#>  1 DOH00000…     NA LIBAC… ""          ""        2020-04-26    NA     NA
#>  2 DOH00000…  34689 LIBAC… "2020-05-0… "2020-05… 2020-05-03    60    250
#>  3 DOH00000…  36433 LIBAC… "2020-05-1… "2020-05… 2020-05-10    60    250
#>  4 DOH00000…  38127 LIBAC… "2020-05-2… "2020-05… 2020-05-17    60    250
#>  5 DOH00000…  39883 LIBAC… "2020-05-3… "2020-05… 2020-05-24    57    280
#>  6 DOH00000… 310866 LIBAC… "2020-06-0… "2020-06… 2020-05-31    57    400
#>  7 DOH00000…     NA LIBAC… ""          ""        2020-06-07    NA     NA
#>  8 DOH00000… 314728 LIBAC… "2020-06-2… "2020-06… 2020-06-14    57    400
#>  9 DOH00000… 316324 LIBAC… "2020-06-2… "2020-06… 2020-06-21    80    240
#> 10 DOH00000…  41172 LIBAC… "2020-07-0… "2020-07… 2020-06-28    57   6500
#> # … with 55,834 more rows, and 18 more variables: head_cover <dbl>,
#> #   goggles <dbl>, coverall <dbl>, shoe_cover <dbl>, face_shield <dbl>,
#> #   surgmask <dbl>, n95mask <dbl>, ops <dbl>, nps <dbl>, vtm <dbl>,
#> #   region <chr>, region_psgc <chr>, province <chr>, province_psgc <chr>,
#> #   city_mun <chr>, city_mun_psgc <chr>, north_coord <dbl>, east_coord <dbl>

## Retrieve latest testing aggregates information
datadrop_get_tests(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 07 Testing Aggregates.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec8492630353c

#> # A tibble: 19,750 x 18
#>    facility_name report_date avg_turnaround_… daily_output_sa… daily_output_un…
#>    <chr>         <chr>                  <dbl>            <dbl>            <dbl>
#>  1 AL Molecular… 2020-08-30                24                0                0
#>  2 AL Molecular… 2020-08-31                24                0                0
#>  3 AL Molecular… 2020-09-01                NA               NA               NA
#>  4 AL Molecular… 2020-09-02                NA               NA               NA
#>  5 AL Molecular… 2020-09-03                NA               NA               NA
#>  6 AL Molecular… 2020-09-04                NA               NA               NA
#>  7 AL Molecular… 2020-09-05                NA               NA               NA
#>  8 AL Molecular… 2020-09-06                NA               NA               NA
#>  9 AL Molecular… 2020-09-07                 8                1                1
#> 10 AL Molecular… 2020-09-08                 8                1                1
#> # … with 19,740 more rows, and 13 more variables:
#> #   daily_output_positive_individuals <dbl>,
#> #   daily_output_negative_individuals <dbl>, daily_output_equivocal <dbl>,
#> #   daily_output_invalid <dbl>, remaining_available_tests <dbl>,
#> #   backlogs <dbl>, cumulative_samples_tested <int>,
#> #   cumulative_unique_individuals <int>, cumulative_positive_individuals <int>,
#> #   cumulative_negative_individuals <int>, pct_positive_cumulative <dbl>,
#> #   pct_negative_cumulative <dbl>, validation_status <chr>

## Retrieve latest daily quarantine facility beds and mechanical ventilators
datadrop_get_qdaily(tbl = x, path = tempfile())

#> Error: The file doesn't seem to have downloaded.

## Retrieve latest weekly quarantine facility PPE and other related equipment
datadrop_get_qweekly(tbl = x, path = tempfile())

#> File downloaded:
#>   * DOH COVID Data Drop_ 20201119 - 09 Quarantine Facility Data - Weekly Report.csv
#> Saved locally as:
#>   * /var/folders/fk/s0yv8hhn2cs_nfsmzhm4dmhc0000gn/T//RtmpDRSiRy/filec849249fa652

#> # A tibble: 1,384 x 25
#>    facilityname updateddate addeddate reportdate coverall surgmask shoe_cover
#>    <chr>        <chr>       <chr>     <chr>         <dbl>    <dbl>      <dbl>
#>  1 GINABLAN IS… ""          ""        2020-06-0…       NA       NA         NA
#>  2 Hijo Elemen… ""          ""        2020-06-0…       NA       NA         NA
#>  3 STA. MARIA … "2020-05-3… "2020-05… 2020-05-3…       56      500         10
#>  4 STA. MARIA … "2020-06-0… "2020-06… 2020-06-0…       55      490         10
#>  5 TAGBINO NAT… ""          ""        2020-06-0…       NA       NA         NA
#>  6 APLAYA SENI… ""          ""        2020-06-0…       NA       NA         NA
#>  7 CAUSWAGAN I… ""          ""        2020-06-0…       NA       NA         NA
#>  8 SAN ISIDRO … ""          ""        2020-06-0…       NA       NA         NA
#>  9 SOLANA SENI… ""          ""        2020-06-0…       NA       NA         NA
#> 10 TAMAYO TRAI… "2020-05-2… "2020-05… 2020-05-2…       56      200        150
#> # … with 1,374 more rows, and 18 more variables: face_shield <dbl>,
#> #   head_cover <dbl>, n95mask <dbl>, goggles <dbl>, gloves <dbl>, gown <dbl>,
#> #   vtm <dbl>, nps <dbl>, ops <dbl>, weekly_reportdate <chr>, region <chr>,
#> #   province <chr>, city_mun <chr>, bgy <chr>, region_psgc <chr>,
#> #   province_psgc <chr>, city_mun_psgc <chr>, bgy_psgc <chr>

#datadrop_get_collectV3(tbl = x, path = tempfile())
#datadrop_get_collectV4(tbl = x, path = tempfile())
#datadrop_get_tracker(tbl = x, path = tempfile())