List files and folders within a specified DoH Data Drop Google Drive folder

List files and folders within a specified DoH Data Drop Google Drive folder

datadrop_ls(id)

Arguments

id
A 33 character string for the Google Drive ID of the latest officially released DoH Data Drop. This can be obtained using the datadrop_latest_id() function or by manually expanding the shortened URL provided in bit.ly/DataDropPH and extracting the Google Drive ID.

Value

A tibble containing information on the various files and folders in the specified Google Drive for DoH Data Drop.

Author

Ernest Guevarra

Examples

id <- datadrop_latest_id()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-13.

datadrop_ls(id = id)

#> # A tibble: 13 x 3
#>    name                                      id                  drive_resource 
#>  * <chr>                                     <chr>               <list>         
#>  1 DOH COVID Data Drop_ 20201113 - 08 Quara… 1FogJUaMIvJnr9Bjmo… <named list [3…
#>  2 DOH COVID Data Drop_ 20201113 - 04 Case … 1CiOgSsDmNGaWPQjCl… <named list [3…
#>  3 DOH COVID Data Drop_ 20201113 - 05 DOH D… 1r_DQste7Mtz17v7iK… <named list [3…
#>  4 DOH COVID Data Drop_ 20201113 - 03 Metad… 1wHXcDkC1c60BrXLSW… <named list [3…
#>  5 DOH Data Drop 20201113 - Changelog.xlsx   1xiiIB7n_vpja6w5PS… <named list [3…
#>  6 DOH COVID Data Drop_ 20201113 - 02 Metad… 1qmgUs5WyhBg6s8MI-… <named list [3…
#>  7 DOH COVID Data Drop_ 20201113 - 09 Quara… 1a4Fi8WlJGE7BOIyAK… <named list [3…
#>  8 DOH COVID Data Drop_ 20201113 - 10 DOH D… 1gan3d0Bn4YoaR9UE8… <named list [3…
#>  9 DOH COVID Data Drop_ 20201113 - 06 DOH D… 1f0dVTkqTz7-zum2ET… <named list [3…
#> 10 DOH COVID Data Drop_ 20201113 - 07 Testi… 1aIBTNadioFSj-OzFO… <named list [3…
#> 11 DOH COVID Data Drop_ 20201113 - 11 DOH D… 1W5NnsO0mmvNDSZEen… <named list [3…
#> 12 DOH COVID Data Drop_ 20201113 - 12 DDC T… 1w4QxKDjc_mHu-X46n… <named list [3…
#> 13 01 READ ME FIRST (11_13).pdf              1W07xG21WHiuP_LbYM… <named list [3…