List files and folders within a specified DoH Data Drop Google Drive folder

List files and folders within a specified DoH Data Drop Google Drive folder

datadrop_ls(id)

Arguments

id
A 33 character string for the Google Drive ID of the latest officially released DoH Data Drop. This can be obtained using the datadrop_id_latest() function or by manually expanding the shortened URL provided in bit.ly/DataDropPH and extracting the Google Drive ID.

Value

A tibble containing information on the various files and folders in the specified Google Drive for DoH Data Drop.

Author

Ernest Guevarra

Examples

## Get Google Drive ID of latest DoH Data Drop
id <- datadrop_id_latest()

#> Getting the Google Drive ID for the latest available DoH Data Drop for
#> 2020-11-18.

#> File downloaded:
#>   * 01 READ ME FIRST (11_18).pdf
#> Saved locally as:
#>   * /var/folders/rx/nr32tl5n6f3d_86tn0tc7kc00000gp/T//RtmpApDPr2/filea2a46f233061

## List the contents of the latest DoH Data Drop in Google Drive
datadrop_ls(id = id)

#> # A tibble: 13 x 3
#>    name                                      id                  drive_resource 
#>  * <chr>                                     <chr>               <list>         
#>  1 DOH COVID Data Drop_ 20201118 - 08 Quara… 1BWlDChxLOJzeehk5s… <named list [3…
#>  2 DOH COVID Data Drop_ 20201118 - 04 Case … 1fYJl0hbiYaq4vPsfb… <named list [3…
#>  3 DOH COVID Data Drop_ 20201118 - 05 DOH D… 18JumMvqHDLvtLzTh0… <named list [3…
#>  4 DOH COVID Data Drop_ 20201118 - 12 DDC T… 1EodO_D2flCenzeqOO… <named list [3…
#>  5 DOH COVID Data Drop_ 20201118 - 11 DOH D… 1e2gfjgqwN604tprFm… <named list [3…
#>  6 DOH Data Drop 20201118 - Changelog.xlsx   1VRHQ38Hwft_bZRCXo… <named list [3…
#>  7 DOH COVID Data Drop_ 20201118 - 10 DOH D… 1vciPP-xK6B5X6RL3O… <named list [3…
#>  8 DOH COVID Data Drop_ 20201118 - 09 Quara… 1TVeE0sP0QQX5BgORA… <named list [3…
#>  9 DOH COVID Data Drop_ 20201118 - 06 DOH D… 1-6DnQTzSc3QP_NGTG… <named list [3…
#> 10 DOH COVID Data Drop_ 20201118 - 02 Metad… 1rJjjtJ0_Dfz6xuxij… <named list [3…
#> 11 DOH COVID Data Drop_ 20201118 - 03 Metad… 1Y0Ilzp2zTwJ7NRZZq… <named list [3…
#> 12 DOH COVID Data Drop_ 20201118 - 07 Testi… 1_WEcJ9NVAREfclVLI… <named list [3…
#> 13 01 READ ME FIRST (11_18).pdf              1jbrSbaE2cLfoaUuH5… <named list [3…